@@ -1,1 +1,1 @@
-Search.setIndex({"docnames": ["api/_autosummary/pyBL", "api/_autosummary/pyBL.blasius", "api/_autosummary/pyBL.blasius.BlasiusSolution", "api/_autosummary/pyBL.dense_output_ppoly", "api/_autosummary/pyBL.falkner_skan", "api/_autosummary/pyBL.falkner_skan.FalknerSkanSolution", "api/_autosummary/pyBL.head_method", "api/_autosummary/pyBL.head_method.HeadMethod", "api/_autosummary/pyBL.ibl_method", "api/_autosummary/pyBL.ibl_method.IBLMethod", "api/_autosummary/pyBL.ibl_method.IBLResult", "api/_autosummary/pyBL.ibl_method.IBLTermEvent", "api/_autosummary/pyBL.initial_condition", "api/_autosummary/pyBL.initial_condition.FalknerSkanStagnationCondition", "api/_autosummary/pyBL.initial_condition.InitialCondition", "api/_autosummary/pyBL.initial_condition.ManualCondition", "api/_autosummary/pyBL.pyBL", "api/_autosummary/pyBL.skin_friction", "api/_autosummary/pyBL.skin_friction.c_f_Felsch", "api/_autosummary/pyBL.skin_friction.c_f_LudwiegTillman", "api/_autosummary/pyBL.skin_friction.c_f_White", "api/_autosummary/pyBL.stanford_olympics", "api/_autosummary/pyBL.stanford_olympics.StanfordOlympics1968", "api/_autosummary/pyBL.thwaites_method", "api/_autosummary/pyBL.thwaites_method.ThwaitesMethod", "api/_autosummary/pyBL.thwaites_method.ThwaitesMethodLinear", "api/_autosummary/pyBL.thwaites_method.ThwaitesMethodNonlinear", "api/_autosummary/pyBL.xfoil_reader", "api/_autosummary/pyBL.xfoil_reader.XFoilReader", "api/modules", "developers/code_of_conduct", "developers/contributing", "developers/index", "examples/index", "index", "license/index", "license/license"], "filenames": ["api/_autosummary/pyBL.rst", "api/_autosummary/pyBL.blasius.rst", "api/_autosummary/pyBL.blasius.BlasiusSolution.rst", "api/_autosummary/pyBL.dense_output_ppoly.rst", "api/_autosummary/pyBL.falkner_skan.rst", "api/_autosummary/pyBL.falkner_skan.FalknerSkanSolution.rst", "api/_autosummary/pyBL.head_method.rst", "api/_autosummary/pyBL.head_method.HeadMethod.rst", "api/_autosummary/pyBL.ibl_method.rst", "api/_autosummary/pyBL.ibl_method.IBLMethod.rst", "api/_autosummary/pyBL.ibl_method.IBLResult.rst", "api/_autosummary/pyBL.ibl_method.IBLTermEvent.rst", "api/_autosummary/pyBL.initial_condition.rst", "api/_autosummary/pyBL.initial_condition.FalknerSkanStagnationCondition.rst", "api/_autosummary/pyBL.initial_condition.InitialCondition.rst", "api/_autosummary/pyBL.initial_condition.ManualCondition.rst", "api/_autosummary/pyBL.pyBL.rst", "api/_autosummary/pyBL.skin_friction.rst", "api/_autosummary/pyBL.skin_friction.c_f_Felsch.rst", "api/_autosummary/pyBL.skin_friction.c_f_LudwiegTillman.rst", "api/_autosummary/pyBL.skin_friction.c_f_White.rst", "api/_autosummary/pyBL.stanford_olympics.rst", "api/_autosummary/pyBL.stanford_olympics.StanfordOlympics1968.rst", "api/_autosummary/pyBL.thwaites_method.rst", "api/_autosummary/pyBL.thwaites_method.ThwaitesMethod.rst", "api/_autosummary/pyBL.thwaites_method.ThwaitesMethodLinear.rst", "api/_autosummary/pyBL.thwaites_method.ThwaitesMethodNonlinear.rst", "api/_autosummary/pyBL.xfoil_reader.rst", "api/_autosummary/pyBL.xfoil_reader.XFoilReader.rst", "api/modules.rst", "developers/code_of_conduct.rst", "developers/contributing.rst", "developers/index.rst", "examples/index.rst", "index.rst", "license/index.rst", "license/license.rst"], "titles": ["pyBL", "pyBL.blasius", "pyBL.blasius.BlasiusSolution", "pyBL.dense_output_ppoly", "pyBL.falkner_skan", "pyBL.falkner_skan.FalknerSkanSolution", "pyBL.head_method", "pyBL.head_method.HeadMethod", "pyBL.ibl_method", "pyBL.ibl_method.IBLMethod", "pyBL.ibl_method.IBLResult", "pyBL.ibl_method.IBLTermEvent", "pyBL.initial_condition", "pyBL.initial_condition.FalknerSkanStagnationCondition", "pyBL.initial_condition.InitialCondition", "pyBL.initial_condition.ManualCondition", "pyBL.pyBL", "pyBL.skin_friction", "pyBL.skin_friction.c_f_Felsch", "pyBL.skin_friction.c_f_LudwiegTillman", "pyBL.skin_friction.c_f_White", "pyBL.stanford_olympics", "pyBL.stanford_olympics.StanfordOlympics1968", "pyBL.thwaites_method", "pyBL.thwaites_method.ThwaitesMethod", "pyBL.thwaites_method.ThwaitesMethodLinear", "pyBL.thwaites_method.ThwaitesMethodNonlinear", "pyBL.xfoil_reader", "pyBL.xfoil_reader.XFoilReader", "API", "Contributor Covenant Code of Conduct", "Contributing", "Developer Information", "Examples", "Summary", "License Information", "GNU General Public License"], "terms": {"python": [0, 34], "implement": [0, 6, 9, 10, 11, 23, 24, 25, 26, 34, 36], "integr": [0, 2, 5, 6, 7, 8, 9, 10, 11, 12, 14, 23, 24, 25, 26, 27, 34], "boundari": [0, 2, 5, 6, 7, 8, 9, 12, 14, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 34], "layer": [0, 2, 5, 6, 7, 8, 9, 12, 14, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 34], "method": [0, 2, 5, 6, 7, 8, 9, 11, 12, 13, 14, 15, 21, 22, 23, 24, 25, 26, 28, 34, 36], "thi": [0, 1, 2, 4, 5, 6, 7, 8, 9, 10, 11, 13, 14, 15, 17, 18, 19, 20, 22, 23, 24, 25, 26, 27, 28, 30, 31, 32, 34, 35], "packag": [0, 10, 11, 36], "i": [0, 1, 2, 4, 5, 7, 9, 10, 11, 14, 18, 19, 20, 21, 22, 23, 24, 25, 26, 28, 30, 31, 32, 34, 35, 36], "an": [0, 2, 8, 9, 11, 22, 24, 25, 26, 28, 30, 31, 34, 36], "modul": [0, 1, 4, 6, 8, 9, 17, 23, 34], "numer": [1, 4, 21], "solut": [1, 2, 4, 5, 7, 9, 12, 13, 14, 24, 25, 26, 27, 34], "equat": [1, 2, 4, 5, 6, 9, 23, 25, 26, 27, 34], "calcul": [1, 2, 4, 5, 7, 9, 17, 18, 19, 20, 22, 24, 25, 26, 34], "laminar": [1, 4, 9, 24, 25, 26, 28, 34], "incompress": [1, 4], "flow": [1, 4, 9, 13, 21, 22, 26, 28, 34, 36], "over": [1, 4], "flat": 1, "plate": 1, "pressur": 1, "gradient": 1, "known": [1, 4, 9], "after": [1, 4, 10, 36], "differenti": [1, 4, 9, 25, 26], "solv": [1, 4, 7, 9, 11, 23, 24, 25, 26, 34], "number": [1, 4, 7, 9, 17, 18, 19, 20, 22, 24, 25, 26, 28, 36], "properti": [1, 2, 4, 5, 7, 9, 22, 24, 25, 26, 34, 36], "can": [1, 2, 4, 5, 7, 9, 11, 12, 24, 25, 26, 31, 34, 35, 36], "obtain": [1, 2, 4, 5, 13, 15, 22, 28, 34], "about": [1, 4, 10, 11, 27, 30, 36], "similar": [1, 2, 4, 5, 36], "coordin": [1, 2, 4, 5, 22, 28], "well": [1, 4, 36], "cartesian": [1, 2, 4, 5], "class": [1, 2, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 21, 22, 23, 24, 25, 26, 27, 28, 34, 36], "u_ref": [2, 5], "nu": [2, 5, 7, 9, 13, 14, 22, 24, 25, 26], "fpp0": 2, "0": [2, 5, 7, 10, 24, 25, 26, 34], "46959998713136886": 2, "eta_inf": [2, 5], "10": [2, 5], "sourc": [2, 5, 7, 9, 10, 11, 13, 14, 15, 18, 19, 20, 22, 24, 25, 26, 28], "base": [2, 5, 7, 8, 9, 10, 11, 12, 13, 14, 15, 18, 19, 20, 22, 23, 24, 25, 26, 28, 34, 36], "object": [2, 5, 9, 10, 22, 24, 25, 26, 28, 36], "repres": [2, 5, 10, 24, 25, 26, 30, 36], "while": [2, 22], "initi": [2, 5, 7, 9, 12, 13, 14, 15, 22, 24, 25, 26, 34, 36], "user": [2, 7, 9, 11, 24, 25, 26], "defin": [2, 30, 36], "paramet": [2, 4, 5, 7, 9, 11, 12, 15, 18, 19, 20, 22, 24, 25, 26, 28, 34], "need": [2, 5, 7, 9, 11, 12, 21, 24, 27, 34, 36], "default": [2, 7, 9, 24, 25, 26, 28], "ar": [2, 4, 7, 8, 9, 10, 11, 18, 19, 20, 21, 22, 23, 24, 25, 26, 28, 30, 31, 32, 34, 36], "suffici": 2, "obain": 2, "accur": [2, 9, 26], "onc": [2, 5], "dens": [2, 5, 9], "output": [2, 5, 9, 36], "from": [2, 5, 7, 8, 9, 10, 11, 12, 13, 15, 17, 18, 19, 20, 22, 24, 25, 26, 28, 30, 31, 34], "od": [2, 5, 7, 9, 23, 24, 25, 26], "us": [2, 5, 7, 9, 11, 12, 18, 19, 20, 21, 22, 24, 25, 26, 28, 30, 31, 32, 34, 35], "report": [2, 5, 22, 28, 30], "back": [2, 5], "wide": [2, 5, 21, 36], "varieti": [2, 5, 21, 22, 34], "associ": [2, 5, 7, 8, 9, 11, 22, 24, 25, 26, 36], "both": [2, 5, 9, 18, 19, 20, 22, 31, 36], "point": [2, 5, 9, 13, 28, 34], "correspond": [2, 5, 18, 19, 20, 22, 34, 36], "d": [2, 5, 7, 9, 24, 25, 26, 34, 36], "x": [2, 5, 7, 9, 10, 22, 24, 25, 26, 28], "rho": [2, 5, 7, 9, 24, 25, 26], "dissip": [2, 5, 7, 9, 24, 25, 26], "arrai": [2, 5, 7, 9, 10, 18, 19, 20, 22, 24, 25, 26, 28], "like": [2, 5, 7, 9, 18, 19, 20, 22, 24, 25, 26, 28, 31, 36], "streamwis": [2, 5, 7, 9, 22, 24, 25, 26], "loation": [2, 5, 7, 9, 24, 25, 26], "float": [2, 5, 7, 9, 10, 11, 13, 14, 15, 18, 19, 20, 22, 24, 25, 26, 28], "freestream": [2, 5, 7, 9, 24, 25, 26, 28], "densiti": [2, 5, 7, 9, 24, 25, 26], "return": [2, 5, 7, 8, 9, 10, 11, 13, 14, 15, 18, 19, 20, 22, 24, 25, 26, 28, 36], "same": [2, 5, 7, 9, 18, 19, 20, 24, 25, 26, 34, 36], "shape": [2, 5, 7, 9, 13, 14, 15, 18, 19, 20, 22, 24, 25, 26, 28, 34], "desir": [2, 5, 7, 9, 22, 24, 25, 26], "specifi": [2, 5, 7, 9, 10, 22, 24, 25, 26, 28, 36], "locat": [2, 5, 7, 9, 10, 22, 24, 25, 26, 28, 31], "h_d": [2, 5, 7, 9, 13, 14, 15, 18, 19, 20, 22, 24, 25, 26, 28], "displac": [2, 5, 7, 9, 13, 14, 15, 17, 18, 19, 20, 22, 24, 25, 26, 28, 34], "factor": [2, 5, 7, 9, 13, 14, 15, 18, 19, 20, 22, 24, 25, 26, 28, 34], "h_k": [2, 5, 7, 9, 13, 14, 15, 22, 24, 25, 26, 28], "kinet": [2, 5, 7, 9, 13, 14, 15, 22, 24, 25, 26, 28], "energi": [2, 5, 7, 9, 13, 14, 15, 22, 24, 25, 26, 28], "u_": [2, 5, 7, 9, 22, 24, 25, 26, 28, 34], "inviscid": [2, 4, 5, 7, 9, 13, 24, 25, 26, 34], "edg": [2, 4, 5, 7, 9, 13, 14, 22, 24, 25, 26, 28, 34], "veloc": [2, 5, 7, 9, 13, 14, 22, 24, 25, 26, 28, 34], "": [2, 5, 6, 7, 9, 10, 22, 24, 25, 26, 28, 30, 36], "interest": [2, 5, 30, 31, 36], "v_e": [2, 5, 7, 9, 24, 25, 26], "transpir": [2, 5, 7, 9, 24, 25, 26, 34], "delta_d": [2, 5, 7, 9, 13, 14, 15, 22, 24, 25, 26, 28], "thick": [2, 5, 7, 9, 13, 14, 15, 17, 18, 19, 20, 22, 24, 25, 26, 28, 34], "delta_k": [2, 5, 7, 9, 13, 14, 15, 22, 24, 25, 26], "delta_m": [2, 5, 7, 9, 13, 14, 15, 22, 24, 25, 26, 28], "momentum": [2, 5, 7, 9, 13, 14, 15, 17, 18, 19, 20, 22, 24, 25, 26, 28, 34], "delta_": [2, 5], "shear": [2, 5, 7, 9, 22, 24, 25, 26, 34], "eta": [2, 5], "y": [2, 5, 22, 28], "llike": [2, 5], "normal": [2, 5, 28, 36], "direct": [2, 5, 9, 36], "either": [2, 5, 11, 22, 34, 35, 36], "note": [2, 5, 11, 22], "must": [2, 5, 36], "eta_d": [2, 5], "eta_k": [2, 5], "eta_m": [2, 5], "eta_": [2, 5], "f": [2, 5, 36], "non": [2, 5, 22], "dimension": [2, 5, 22], "stream": [2, 5], "function": [2, 5, 7, 9, 17, 18, 19, 20, 24, 25, 26, 36], "valu": [2, 5, 7, 9, 10, 11], "fp": [2, 5], "fpp": [2, 5], "deriv": [2, 5, 7, 9, 11, 24, 25, 26], "tau_w": [2, 5, 7, 9, 24, 25, 26, 34], "wall": [2, 5, 7, 9, 22, 24, 25, 26, 34], "stress": [2, 5, 7, 9, 24, 25, 26, 34], "u": [2, 5, 31], "compon": [2, 5, 36], "v": [2, 5], "creat": [3, 9, 12, 30], "thu": [3, 36], "aug": 3, "11": 3, "15": 3, "59": 3, "40": 3, "2022": [3, 24, 25, 26], "author": [3, 36], "ddmarshal": 3, "falkner": [4, 5, 13], "skan": [4, 5, 13], "wedg": 4, "These": [4, 9, 12, 22, 31, 34], "depend": [4, 9], "veociti": 4, "profil": [4, 7, 14, 25, 26], "m": [4, 5, 22, 28, 34], "which": [4, 30, 34, 36], "relat": [4, 7, 9, 11, 18, 19, 20, 24, 25, 26], "angl": [4, 28], "It": [5, 24, 34, 36], "perform": [5, 36], "search": [5, 34], "appropri": [5, 30, 36], "condit": [5, 7, 9, 12, 13, 14, 15, 22, 24, 25, 26], "dure": 5, "head": [6, 7, 34], "contain": [6, 22, 23, 27, 28, 31, 36], "necessari": [6, 23, 30, 36], "data": [6, 8, 12, 18, 19, 20, 21, 22, 23, 24, 25, 26, 28, 36], "two": [6, 11, 21, 23, 36], "1": [7, 10, 11, 22, 24, 25, 26, 28, 30, 34], "none": [7, 9, 22, 24, 25, 26, 28], "du_edx": [7, 9, 13, 14, 22, 24, 25, 26], "d2u_edx2": [7, 9, 24, 25, 26], "h_d_crit": 7, "2": [7, 9, 22, 24, 25, 26, 28, 34], "4": [7, 30, 34], "iblmethod": [7, 8, 10, 11, 12, 24, 25, 26], "model": [7, 24, 25, 26, 28, 34, 36], "turbul": [7, 17, 18, 19, 20, 21, 22, 28, 34], "bondari": 7, "1958": [7, 22, 34], "system": [7, 36], "when": [7, 8, 9, 11, 22, 24, 25, 26, 28, 30, 31, 36], "provid": [7, 8, 9, 10, 11, 12, 14, 15, 17, 22, 23, 24, 25, 26, 30, 31, 34, 36], "other": [7, 9, 10, 11, 27, 30, 34], "configur": [7, 9, 24, 25, 26, 34], "inform": [7, 9, 10, 11, 21, 22, 24, 25, 26, 27, 30, 31, 34, 36], "rais": [7, 9, 22, 24, 25, 26, 28], "typeerror": [7, 9, 24, 25, 26], "have": [7, 9, 11, 18, 19, 20, 24, 25, 26, 30, 31, 34], "been": [7, 9, 24, 25, 26, 36], "set": [7, 9, 10, 12, 15, 21, 24, 25, 26, 30, 34], "second": [7, 9, 24, 25, 26], "kinemat": [7, 9, 13, 14, 22, 24, 25, 26], "viscos": [7, 9, 13, 14, 22, 24, 25, 26], "set_h_d_crit": 7, "critic": [7, 30], "separ": [7, 10, 11, 28, 30, 36], "sinc": 7, "doe": [7, 36], "predict": [7, 21], "skin": [7, 17, 18, 19, 20, 22, 28], "friction": [7, 17, 18, 19, 20, 22, 28], "zero": 7, "anoth": 7, "mechan": 7, "emploi": 7, "determin": [7, 14, 30, 36], "occur": [7, 10, 36], "threshold": 7, "indic": [7, 11, 22, 36], "ha": [7, 8, 9, 10, 11, 36], "new": [7, 28], "set_initial_condit": [7, 9, 24, 25, 26], "ic": [7, 9, 24, 25, 26], "initialcondit": [7, 9, 12, 13, 15, 24, 25, 26], "solver": [7, 8, 9, 10, 11, 24, 25, 26], "set_initial_paramet": [7, 24, 25, 26], "delta_m0": [7, 24, 25, 26, 34], "h_d0": [7, 34], "propag": [7, 36], "start": [7, 9, 12, 24, 25, 26, 36], "valueerror": [7, 9, 24, 25, 26], "neg": [7, 24, 25, 26], "invalid": [7, 9, 22, 24, 25, 26, 36], "set_veloc": [7, 9, 24, 25, 26], "There": [7, 8, 9, 23, 24, 25, 26], "differ": [7, 9, 24, 25, 26, 30, 36], "wai": [7, 9, 24, 25, 26, 36], "its": [7, 9, 24, 25, 26, 27, 30, 36], "see": [7, 9, 11, 22, 24, 25, 26, 30, 31, 34, 35, 36], "definit": [7, 9, 22, 24, 25, 26], "detail": [7, 9, 10, 11, 24, 25, 26, 30, 34, 35, 36], "tupl": [7, 9, 22, 24, 25, 26, 28], "scalar": [7, 9, 24, 25, 26], "represent": [7, 9, 22, 24, 25, 26, 30], "analysi": [7, 9, 24, 25, 26, 27], "option": [7, 9, 24, 25, 26, 28, 34, 35, 36], "first": [7, 9, 24, 25, 26, 36], "The": [7, 9, 10, 11, 21, 22, 24, 25, 26, 28, 30, 31, 34, 36], "representationa": [7, 9, 24, 25, 26], "messag": [7, 9, 10, 24, 25, 26, 31, 34], "x0": [7, 9, 24, 25, 26, 34], "x_end": [7, 9, 10, 24, 25, 26, 34], "term_ev": [7, 9, 24, 25, 26], "iblresult": [7, 8, 9, 24, 25, 26], "thwait": [23, 24, 25, 26, 34], "actual": 36, "follow": [24, 25, 26, 30, 31, 36], "frac": [25, 26], "dx": [25, 26], "left": [25, 26], "right": [25, 26, 30], "where": 36, "linear": [23, 24, 25, 26], "approxim": [9, 24, 25, 26, 34, 36], "term": [22, 34, 35], "origin": [22, 24, 25, 26, 34, 36], "paper": [25, 36], "end": [7, 9, 10, 24, 25, 26, 36], "list": [7, 9, 24, 25, 26, 36], "ibltermev": [7, 9, 24, 25, 26], "event": [7, 9, 11, 24, 25, 26, 30, 36], "termin": [7, 9, 10, 11, 24, 25, 26], "process": [7, 9, 10, 24, 25, 26, 28], "befor": [7, 9, 24, 25, 26, 31], "reach": [7, 9, 10, 24, 25, 26], "all": [8, 9, 22, 23, 27, 30, 36], "lot": 8, "common": [8, 9, 12, 30, 31, 36], "characterst": 8, "variou": 8, "much": [8, 10, 31], "code": [8, 32, 34], "possibl": [8, 10, 31, 36], "should": [8, 11, 31, 34, 36], "inherit": 8, "instanc": [8, 30], "complet": [7, 8, 9, 10, 24, 25, 26], "abc": [9, 11, 14], "encapsul": 9, "featur": [9, 34, 36], "ibl": [7, 9, 10, 11, 12, 24, 25, 26, 31, 34], "At": 9, "veri": 9, "least": [9, 36], "intefac": 9, "expect": [9, 30, 36], "intengr": 9, "one": [9, 23, 24, 25, 26, 36], "more": [9, 26, 32, 34, 35, 36], "ordinari": 9, "explicitli": [9, 36], "solve_ivp": [9, 10, 11], "scipi": [9, 10], "store": [9, 22, 27, 28], "result": [9, 10, 23, 30, 36], "interpol": 9, "polynomi": 9, "degre": 9, "allow": [9, 36], "queri": [9, 34], "ani": [9, 11, 30, 34, 35, 36], "between": [9, 22, 34, 36], "uniformli": 9, "In": [9, 24, 30, 31, 34, 36], "order": [9, 12, 30, 36], "variat": 9, "sedond": 9, "xpoint": 9, "case": [9, 22, 23, 27, 28, 36], "monoton": 9, "cubic": 9, "spline": [9, 24, 25, 26], "taken": 9, "rate": [9, 13, 22], "chang": [9, 11, 13, 22, 34, 36], "found": [9, 36], "antideriv": 9, "pass": [9, 36], "callabl": [9, 24, 25, 26], "If": [9, 18, 19, 20, 31, 34, 36], "call": [9, 11, 36], "gener": [9, 11, 27, 34, 35], "otherwis": [9, 30, 34, 36], "finit": [9, 24, 25, 26], "neither": [9, 36], "mai": [9, 11, 30, 36], "specif": [7, 9, 11, 24, 25, 26, 30, 36], "being": [9, 36], "By": [9, 36], "assum": [9, 36], "stagnat": [9, 13, 15, 28], "abstract": [9, 11, 12, 14], "inf": [10, 28], "f_end": 10, "statu": [10, 30, 31, 36], "99": 10, "Not": 10, "success": [10, 34], "fals": 10, "bunch": 10, "within": [10, 11, 30, 36], "To": [10, 31, 36], "most": [10, 31, 36], "import": [10, 34, 36], "attribut": [10, 13, 14, 15, 22, 28, 36], "np": [10, 28], "state": [10, 36], "int": [10, 22, 28], "reason": [10, 11, 30, 36], "final": [10, 36], "distanc": [10, 22, 28], "transit": [10, 11, 28, 34], "string": [10, 22, 24, 25, 26, 28], "descript": [10, 11], "boolean": 10, "true": 10, "successfulli": 10, "event_info": 11, "_call_impl": 11, "valid": [11, 36], "e": [11, 30, 36], "why": [11, 36], "particular": [7, 9, 11, 22, 24, 25, 26, 34, 35, 36], "sign": [11, 36], "document": [11, 32, 36], "purpos": [11, 34, 35, 36], "feedback": 11, "what": [11, 30, 31, 36], "caus": [11, 36], "help": [11, 32], "index": [11, 28, 34], "would": [11, 31, 36], "extra": 11, "consid": [11, 30, 36], "situat": 11, "cleanli": 11, "handl": 11, "out": [12, 36], "interfac": [12, 14, 22, 28, 34, 36], "custom": [12, 36], "thei": [12, 18, 19, 20, 30, 31, 36], "inher": 12, "intend": [14, 34, 36], "due_dx": 14, "slope": [14, 24, 25, 26], "manual": 15, "del_d": 15, "del_m": 15, "del_k": 15, "coeffici": [17, 18, 19, 20, 22, 28], "reynold": [17, 18, 19, 20, 22, 28], "re_delta_m": [18, 19, 20, 22], "felsch": 18, "1968": [18, 21, 22], "given": [18, 19, 20, 34, 36], "experiment": [18, 19, 20, 21, 22], "input": [18, 19, 20], "ludwieg": [19, 22], "tillman": [19, 22], "1950": 19, "white": [20, 24, 25, 26], "2011": [20, 24, 25, 26], "stanford": [21, 22], "olymp": [21, 22], "name": [21, 24, 25, 26, 28, 36], "comput": [21, 22, 36], "afosr": [21, 22], "ifp": [21, 22], "confer": [21, 22], "met": [21, 36], "proceed": [21, 22], "volum": [21, 22, 36], "dedic": 21, "behavior": [21, 30], "ii": [21, 22], "currat": 21, "collect": [21, 22, 36], "high": 21, "qualiti": [21, 32, 36], "also": [22, 30, 31, 36], "refer": [22, 36], "come": [22, 36], "compil": [22, 36], "were": [22, 36], "consist": [22, 36], "measur": [22, 36], "station": [22, 28], "along": [22, 24, 34, 36], "surfac": [22, 28, 34], "throughout": 22, "mixtur": 22, "english": 22, "si": 22, "unit": [22, 31], "page": [22, 34], "ix": 22, "precis": [22, 36], "each": [22, 28, 36], "four": 22, "digit": 22, "g": 22, "equilibrium": 22, "stationdata": 22, "row": [22, 28], "si_unit": 22, "summari": 22, "tabl": 22, "flag": 22, "whether": [22, 36], "c_f": [22, 28], "c_f_lt": 22, "formula": 22, "c_f_e": 22, "beta": 22, "u_star": 22, "delta_c": 22, "clauser": 22, "re_delta_d": 22, "y_plu": 22, "u_plu": 22, "local": [22, 36], "u_on_u": 22, "rel": [22, 34], "y_on_delta_c": 22, "udef": 22, "defect": [22, 28, 36], "change_case_data": [22, 28], "reset": [22, 28], "load": 22, "file": [22, 27, 28, 36], "librari": [22, 32, 34, 36], "except": [22, 28, 36], "read": [22, 28, 36], "num_stat": 22, "want": [22, 36], "entir": [22, 36], "3": [22, 24, 25, 26, 34, 35], "posit": [22, 30], "velocity_smooth": 22, "smooth": 22, "concret": [23, 34], "thwaitesmethodlinear": [23, 24], "tradit": 23, "assumpt": [23, 36], "fit": [23, 24, 25, 26, 34, 35, 36], "relationship": [23, 36], "thwaitesmethodnonlinear": [23, 24, 34], "remov": [23, 30, 36], "slightli": 23, "better": 23, "test": 23, "data_fit": [24, 25, 26], "aeronaut": [24, 34], "journal": [24, 34], "vol": [24, 34], "No": [24, 34], "1949": [24, 26, 34], "pp": [24, 34], "245": [24, 34], "280": [24, 34], "nonlinear": 24, "version": [24, 30, 35], "addit": [24, 31, 34], "algorithm": 24, "reli": [24, 36], "upon": 24, "tabul": 24, "few": [24, 25, 26, 34], "set_data_fit": [24, 25, 26], "via": [24, 25, 26, 30, 31], "take": [24, 25, 26, 27, 30, 34, 36], "three": [24, 25, 26, 36], "intern": [24, 25, 26], "edland": [24, 25, 26, 34], "curv": [24, 25, 26], "cebeci": [24, 25, 26], "bradshaw": [24, 25, 26], "1977": [24, 25, 26], "unus": [24, 25, 26], "thwaitesmethod": [25, 26], "tabular": [25, 26], "work": [25, 26, 31, 34, 36], "45": 25, "6": 25, "lambda": [25, 26], "exact": [26, 36], "without": [26, 30, 34, 35, 36], "math": 26, "h": 26, "captur": 26, "pars": 27, "xfoil": [27, 28], "dump": [27, 28], "character": [27, 36], "filenam": [27, 28], "comparison": 27, "airfoil": 28, "alpha": 28, "c": [28, 36], "re": 28, "x_tran": 28, "n_tran": 28, "reader": 28, "wake": 28, "upper": 28, "lower": 28, "lead": 28, "portion": [28, 36], "analyz": [28, 31], "attack": [28, 30], "chord": 28, "amplif": 28, "length": 28, "airfoildata": 28, "arclength": 28, "nondimension": 28, "p": 28, "mass": 28, "k": 28, "h_d_lower": 28, "h_d_upper": 28, "h_d_wake": 28, "h_k_lower": 28, "h_k_upper": 28, "k_lower": 28, "k_upper": 28, "p_lower": 28, "p_upper": 28, "u_e_low": 28, "u_e_upp": 28, "u_e_wak": 28, "wakedata": 28, "c_f_lower": 28, "c_f_upper": 28, "correctli": 28, "format": [28, 36], "delta_d_low": 28, "delta_d_upp": 28, "delta_d_wak": 28, "delta_k_low": 28, "delta_k_upp": 28, "delta_m_low": 28, "rrai": 28, "delta_m_upp": 28, "delta_m_wak": 28, "m_lower": 28, "m_upper": 28, "num_points_low": 28, "airofil": 28, "num_points_upp": 28, "num_points_wak": 28, "point_low": 28, "point_upp": 28, "point_wak": 28, "s_lower": 28, "arc": 28, "s_upper": 28, "s_wake": 28, "trail": 28, "x_lower": 28, "x_upper": 28, "x_wake": 28, "y_lower": 28, "y_upper": 28, "y_wake": 28, "we": [30, 31, 36], "member": 30, "commun": [30, 31, 36], "healthi": 30, "welcom": [30, 31, 36], "environ": [30, 31], "everyon": [30, 31, 36], "foster": [30, 31], "open": [30, 31], "maintain": [30, 31, 36], "make": [30, 31, 36], "particip": [30, 31], "project": [30, 31, 34], "harass": [30, 31], "free": [30, 31, 34, 35, 36], "experi": [30, 31], "regardless": [30, 31, 36], "ag": [30, 31], "bodi": [30, 31, 34, 36], "size": [30, 31], "disabl": [30, 31], "ethnic": [30, 31], "sex": [30, 31], "characterist": [30, 31], "gender": [30, 31], "ident": [30, 31], "express": [30, 31, 36], "level": [30, 31], "educ": [30, 31], "socio": [30, 31], "econom": [30, 31], "nation": [30, 31], "person": [30, 31, 36], "appear": [30, 31], "race": [30, 31], "religion": [30, 31], "sexual": [30, 31], "orient": [30, 31], "exampl": [30, 31, 36], "contribut": [30, 32, 34], "includ": [30, 36], "inclus": [30, 36], "languag": [30, 36], "Being": 30, "respect": [30, 36], "viewpoint": 30, "gracefulli": 30, "accept": 30, "construct": 30, "focus": 30, "best": [30, 31, 36], "show": [30, 36], "empathi": 30, "toward": 30, "unaccept": [30, 36], "imageri": 30, "unwelcom": 30, "attent": 30, "advanc": 30, "troll": 30, "insult": 30, "derogatori": 30, "comment": 30, "polit": 30, "public": [30, 34, 35], "privat": [30, 36], "publish": [30, 34, 35, 36], "physic": [30, 36], "electron": [30, 36], "address": [30, 36], "explicit": 30, "permiss": 30, "could": [30, 34, 36], "inappropri": 30, "profession": 30, "clarifi": 30, "fair": [30, 36], "correct": [30, 36], "action": [30, 36], "edit": 30, "reject": 30, "commit": [30, 31, 36], "wiki": 30, "issu": [30, 31], "align": 30, "ban": 30, "temporarili": 30, "perman": [30, 36], "deem": [30, 36], "threaten": [30, 36], "offens": 30, "harm": 30, "appli": 30, "space": 30, "individu": [30, 36], "offici": [30, 36], "mail": [30, 36], "post": 30, "social": 30, "media": 30, "account": 30, "act": 30, "appoint": 30, "onlin": 30, "offlin": 30, "further": [30, 36], "abus": [30, 36], "contact": [30, 36], "team": 30, "complaint": 30, "review": [30, 36], "investig": 30, "circumst": [30, 36], "oblig": [30, 36], "confidenti": 30, "regard": [30, 36], "incid": 30, "polici": 30, "who": [30, 32, 36], "do": [30, 31, 32, 36], "good": 30, "faith": 30, "face": 30, "temporari": 30, "repercuss": 30, "leadership": 30, "adapt": [30, 36], "avail": [30, 31, 36], "http": [30, 34, 36], "www": [30, 34, 36], "org": [30, 34, 36], "html": [30, 36], "For": [30, 31, 34, 36], "answer": 30, "question": 30, "faq": 30, "main": [31, 34], "repositori": 31, "github": [31, 34], "you": [31, 34, 35, 36], "let": 31, "know": [31, 36], "contributor": [31, 32, 36], "pledg": 31, "our": [31, 36], "root": 31, "directori": 31, "tbd": 31, "pep": 31, "8": 31, "produc": [31, 36], "warn": 31, "pycodestyl": 31, "pylint": 31, "check": 31, "id": 31, "support": [31, 34, 36], "tool": [31, 36], "instal": [31, 36], "pypi": 31, "pleas": [31, 36], "discuss": 31, "your": [31, 34, 35], "plan": 31, "done": 31, "email": 31, "tracker": 31, "reproduc": 31, "error": 31, "A": [31, 34, 35, 36], "minim": 31, "easier": 31, "track": 31, "down": 31, "problem": [31, 36], "those": [32, 36], "interst": 32, "capabl": 32, "improv": 32, "anyth": [32, 36], "coven": [32, 34, 36], "conduct": [32, 34], "viscou": 34, "effect": [34, 36], "thin": 34, "current": 34, "One": 34, "thwaites1949": 34, "singl": [34, 36], "distribut": [34, 35, 36], "thwaites_method": 34, "rho_inf": 34, "nu_inf": 34, "codit": 34, "tm": 34, "rtn": 34, "print": 34, "get": [34, 36], "els": 34, "tau_wal": 34, "similarli": 34, "head1958": 34, "head_method": 34, "headmethod": 34, "hm": 34, "b": [34, 36], "r": 34, "entrain": 34, "3152": 34, "ministri": 34, "aviat": 34, "research": 34, "council": 34, "david": 34, "marshal": 34, "malachi": 34, "michel": 34, "criteria": 34, "5": 34, "updat": [34, 36], "simplifi": 34, "ad": [34, 36], "clean": 34, "up": [7, 9, 24, 25, 26, 34], "last": 34, "releas": [34, 36], "directli": [34, 36], "thesi": 34, "mostli": 34, "program": [34, 35], "softwar": [34, 35, 36], "redistribut": [34, 35, 36], "modifi": [34, 35], "under": [34, 35, 36], "gnu": [34, 35], "foundat": [34, 35, 36], "later": [34, 35, 36], "hope": [34, 35, 36], "warranti": [34, 35], "even": [34, 35, 36], "impli": [34, 35, 36], "merchant": [34, 35, 36], "FOR": [34, 35, 36], "receiv": [34, 36], "copi": 34, "attach": [34, 36], "form": 34, "some": [34, 36], "develop": [34, 36], "api": 34, "might": [34, 36], "home": 34, "resourc": 34, "29": 36, "june": 36, "2007": 36, "copyright": 36, "inc": 36, "fsf": 36, "permit": 36, "copyleft": 36, "kind": 36, "practic": 36, "design": 36, "awai": 36, "share": 36, "contrast": 36, "guarante": 36, "sure": 36, "remain": 36, "too": 36, "speak": 36, "price": 36, "charg": 36, "them": 36, "wish": 36, "piec": 36, "thing": 36, "prevent": 36, "deni": 36, "ask": 36, "therefor": 36, "certain": 36, "respons": 36, "grati": 36, "fee": 36, "And": 36, "so": 36, "gpl": 36, "step": 36, "assert": 36, "offer": 36, "give": 36, "clearli": 36, "explain": 36, "sake": 36, "mark": 36, "erron": 36, "previou": 36, "devic": 36, "access": 36, "run": [7, 9, 24, 25, 26, 36], "insid": 36, "although": 36, "manufactur": 36, "fundament": 36, "incompat": 36, "aim": 36, "systemat": 36, "pattern": 36, "area": 36, "product": 36, "prohibit": 36, "aris": 36, "substanti": 36, "domain": 36, "stand": 36, "readi": 36, "extend": 36, "provis": 36, "futur": 36, "everi": 36, "constantli": 36, "restrict": 36, "avoid": 36, "special": 36, "danger": 36, "proprietari": 36, "assur": 36, "cannot": 36, "render": 36, "modif": 36, "mean": 36, "semiconductor": 36, "mask": 36, "license": 36, "organ": 36, "part": 36, "fashion": 36, "than": 36, "earlier": 36, "cover": 36, "unmodifi": 36, "secondarili": 36, "liabl": 36, "infring": 36, "applic": 36, "execut": 36, "countri": 36, "activ": 36, "enabl": 36, "parti": 36, "mere": 36, "interact": 36, "through": 36, "network": 36, "transfer": 36, "displai": [34, 36], "notic": 36, "extent": 36, "conveni": 36, "promin": 36, "visibl": 36, "tell": 36, "view": 36, "present": 36, "command": 36, "menu": 36, "item": 36, "meet": 36, "criterion": 36, "prefer": 36, "standard": 36, "recogn": 36, "among": 36, "whole": 36, "major": 36, "serv": 36, "onli": 36, "context": 36, "essenti": 36, "kernel": 36, "window": 36, "oper": 36, "script": 36, "control": 36, "howev": 36, "dynam": 36, "link": 36, "subprogram": 36, "intim": 36, "regener": 36, "grant": 36, "irrevoc": 36, "affirm": 36, "unlimit": 36, "content": 36, "constitut": 36, "acknowledg": 36, "equival": 36, "long": 36, "forc": 36, "sole": 36, "exclus": 36, "facil": 36, "compli": 36, "materi": 36, "behalf": 36, "outsid": 36, "below": 36, "sublicens": 36, "unnecessari": 36, "shall": 36, "technolog": 36, "fulfil": 36, "articl": 36, "wipo": 36, "treati": 36, "adopt": 36, "20": 36, "decemb": 36, "1996": 36, "waiv": 36, "power": 36, "forbid": 36, "exercis": 36, "intent": 36, "enforc": 36, "against": 36, "third": 36, "medium": 36, "conspicu": 36, "keep": 36, "intact": 36, "accord": 36, "absenc": 36, "carri": 36, "relev": 36, "date": 36, "anyon": 36, "possess": 36, "independ": 36, "natur": 36, "extens": 36, "combin": 36, "larger": 36, "storag": 36, "aggreg": 36, "beyond": 36, "machin": 36, "readabl": 36, "embodi": 36, "accompani": 36, "fix": [34, 36], "durabl": 36, "customarili": 36, "interchang": 36, "written": 36, "year": 36, "spare": 36, "cost": 36, "server": 36, "altern": 36, "occasion": 36, "noncommerci": 36, "subsect": 36, "6b": 36, "place": 36, "clear": 36, "next": 36, "sai": 36, "find": 36, "host": 36, "ensur": 36, "satisfi": 36, "peer": 36, "transmiss": 36, "6d": 36, "whose": 36, "exclud": 36, "consum": 36, "tangibl": 36, "famili": 36, "household": 36, "sold": 36, "incorpor": 36, "dwell": 36, "doubt": 36, "resolv": 36, "favor": 36, "coverag": 36, "typic": 36, "commerci": 36, "industri": 36, "unless": 36, "signific": 36, "mode": 36, "procedur": 36, "kei": 36, "suffic": 36, "continu": 36, "interf": 36, "becaus": [7, 9, 24, 25, 26, 36], "made": 36, "transact": 36, "perpetu": 36, "But": 36, "nor": 36, "retain": 36, "abil": 36, "rom": 36, "servic": 36, "itself": 36, "advers": 36, "affect": 36, "violat": 36, "rule": 36, "protocol": 36, "across": 36, "publicli": 36, "password": 36, "unpack": 36, "supplement": 36, "treat": 36, "though": 36, "govern": 36, "own": 36, "notwithstand": 36, "add": 36, "holder": 36, "preserv": 36, "misrepresent": 36, "licensor": 36, "declin": 36, "trademark": 36, "trade": 36, "indemnif": 36, "contractu": 36, "impos": 36, "relicens": 36, "surviv": 36, "statement": 36, "abov": 36, "expressli": 36, "attempt": 36, "void": 36, "paragraph": 36, "ceas": 36, "reinstat": 36, "provision": 36, "until": 36, "fail": 36, "notifi": 36, "prior": 36, "60": 36, "dai": 36, "cessat": 36, "moreov": 36, "time": 36, "cure": 36, "30": 36, "receipt": 36, "qualifi": 36, "ancillari": 36, "consequ": 36, "likewis": 36, "noth": 36, "subject": 36, "complianc": 36, "entiti": 36, "asset": 36, "subdivid": 36, "merg": 36, "whatev": 36, "predecessor": 36, "had": 36, "plu": 36, "effort": 36, "royalti": 36, "litig": 36, "cross": 36, "claim": 36, "counterclaim": 36, "lawsuit": 36, "alleg": 36, "sell": 36, "sale": 36, "alreadi": 36, "acquir": 36, "hereaft": 36, "manner": 36, "worldwid": 36, "agreement": 36, "denomin": 36, "sue": 36, "knowingli": 36, "readili": 36, "arrang": 36, "depriv": 36, "yourself": 36, "benefit": 36, "knowledg": 36, "identifi": 36, "believ": 36, "pursuant": 36, "connect": 36, "procur": 36, "convey": 36, "discriminatori": 36, "scope": 36, "busi": 36, "payment": 36, "primarili": 36, "enter": 36, "wa": [7, 9, 24, 25, 26, 36], "28": 36, "march": 36, "constru": 36, "defens": 36, "court": 36, "contradict": 36, "excus": 36, "simultan": 36, "pertin": 36, "agre": 36, "whom": 36, "refrain": 36, "concern": 36, "Such": 36, "spirit": 36, "distinguish": 36, "choos": 36, "ever": 36, "proxi": 36, "decid": 36, "THERE": 36, "NO": 36, "THE": 36, "TO": 36, "BY": 36, "IN": 36, "write": 36, "OR": 36, "AS": 36, "OF": 36, "BUT": 36, "NOT": 36, "risk": 36, "WITH": 36, "prove": 36, "repair": 36, "WILL": 36, "BE": 36, "damag": 36, "incident": 36, "consequenti": 36, "inabl": 36, "loss": 36, "BEING": 36, "inaccur": 36, "sustain": 36, "failur": 36, "IF": 36, "SUCH": 36, "advis": 36, "close": 36, "absolut": 36, "waiver": 36, "civil": 36, "greatest": 36, "achiev": 36, "safest": 36, "line": 36, "pointer": 36, "full": 36, "brief": 36, "idea": 36, "short": 36, "type": 36, "w": 36, "hypothet": 36, "Of": 36, "cours": 36, "gui": 36, "box": 36, "employ": 36, "programm": 36, "school": 36, "subroutin": 36, "lesser": 36, "instead": 36, "philosophi": 36, "lgpl": 36, "child": [7, 9, 24, 25, 26], "trigger": [7, 9, 24, 25, 26], "minor": 34}, "objects": {"": [[0, 0, 0, "-", "pyBL"]], "pyBL": [[1, 0, 0, "-", "blasius"], [3, 0, 0, "-", "dense_output_ppoly"], [4, 0, 0, "-", "falkner_skan"], [6, 0, 0, "-", "head_method"], [8, 0, 0, "-", "ibl_method"], [12, 0, 0, "-", "initial_condition"], [16, 0, 0, "-", "pyBL"], [17, 0, 0, "-", "skin_friction"], [21, 0, 0, "-", "stanford_olympics"], [23, 0, 0, "-", "thwaites_method"], [27, 0, 0, "-", "xfoil_reader"]], "pyBL.blasius": [[2, 1, 1, "", "BlasiusSolution"]], "pyBL.blasius.BlasiusSolution": [[2, 2, 1, "", "D"], [2, 2, 1, "", "H_d"], [2, 2, 1, "", "H_k"], [2, 2, 1, "", "U_e"], [2, 2, 1, "", "V_e"], [2, 2, 1, "", "delta_d"], [2, 2, 1, "", "delta_k"], [2, 2, 1, "", "delta_m"], [2, 2, 1, "", "delta_s"], [2, 2, 1, "", "eta"], [2, 2, 1, "", "eta_d"], [2, 2, 1, "", "eta_k"], [2, 2, 1, "", "eta_m"], [2, 2, 1, "", "eta_s"], [2, 2, 1, "", "f"], [2, 2, 1, "", "fp"], [2, 2, 1, "", "fpp"], [2, 2, 1, "", "tau_w"], [2, 2, 1, "", "u"], [2, 2, 1, "", "v"]], "pyBL.falkner_skan": [[5, 1, 1, "", "FalknerSkanSolution"]], "pyBL.falkner_skan.FalknerSkanSolution": [[5, 2, 1, "", "D"], [5, 2, 1, "", "H_d"], [5, 2, 1, "", "H_k"], [5, 2, 1, "", "U_e"], [5, 2, 1, "", "V_e"], [5, 2, 1, "", "delta_d"], [5, 2, 1, "", "delta_k"], [5, 2, 1, "", "delta_m"], [5, 2, 1, "", "delta_s"], [5, 2, 1, "", "eta"], [5, 2, 1, "", "eta_d"], [5, 2, 1, "", "eta_k"], [5, 2, 1, "", "eta_m"], [5, 2, 1, "", "eta_s"], [5, 2, 1, "", "f"], [5, 2, 1, "", "fp"], [5, 2, 1, "", "fpp"], [5, 2, 1, "", "tau_w"], [5, 2, 1, "", "u"], [5, 2, 1, "", "v"]], "pyBL.head_method": [[7, 1, 1, "", "HeadMethod"]], "pyBL.head_method.HeadMethod": [[7, 2, 1, "", "D"], [7, 2, 1, "", "H_d"], [7, 2, 1, "", "H_k"], [7, 2, 1, "", "U_e"], [7, 2, 1, "", "V_e"], [7, 2, 1, "", "d2U_edx2"], [7, 2, 1, "", "dU_edx"], [7, 2, 1, "", "delta_d"], [7, 2, 1, "", "delta_k"], [7, 2, 1, "", "delta_m"], [7, 2, 1, "", "nu"], [7, 2, 1, "", "set_H_d_critical"], [7, 2, 1, "", "set_initial_condition"], [7, 2, 1, "", "set_initial_parameters"], [7, 2, 1, "", "set_velocity"], [7, 2, 1, "", "solve"], [7, 2, 1, "", "tau_w"]], "pyBL.ibl_method": [[9, 1, 1, "", "IBLMethod"], [10, 1, 1, "", "IBLResult"], [11, 1, 1, "", "IBLTermEvent"]], "pyBL.ibl_method.IBLMethod": [[9, 2, 1, "", "D"], [9, 2, 1, "", "H_d"], [9, 2, 1, "", "H_k"], [9, 2, 1, "", "U_e"], [9, 2, 1, "", "V_e"], [9, 2, 1, "", "d2U_edx2"], [9, 2, 1, "", "dU_edx"], [9, 2, 1, "", "delta_d"], [9, 2, 1, "", "delta_k"], [9, 2, 1, "", "delta_m"], [9, 2, 1, "", "nu"], [9, 2, 1, "", "set_initial_condition"], [9, 2, 1, "", "set_velocity"], [9, 2, 1, "", "solve"], [9, 2, 1, "", "tau_w"]], "pyBL.ibl_method.IBLTermEvent": [[11, 2, 1, "", "event_info"]], "pyBL.initial_condition": [[13, 1, 1, "", "FalknerSkanStagnationCondition"], [14, 1, 1, "", "InitialCondition"], [15, 1, 1, "", "ManualCondition"]], "pyBL.initial_condition.FalknerSkanStagnationCondition": [[13, 2, 1, "", "H_d"], [13, 2, 1, "", "H_k"], [13, 2, 1, "", "delta_d"], [13, 2, 1, "", "delta_k"], [13, 2, 1, "", "delta_m"]], "pyBL.initial_condition.InitialCondition": [[14, 2, 1, "", "H_d"], [14, 2, 1, "", "H_k"], [14, 2, 1, "", "delta_d"], [14, 2, 1, "", "delta_k"], [14, 2, 1, "", "delta_m"]], "pyBL.initial_condition.ManualCondition": [[15, 2, 1, "", "H_d"], [15, 2, 1, "", "H_k"], [15, 2, 1, "", "delta_d"], [15, 2, 1, "", "delta_k"], [15, 2, 1, "", "delta_m"]], "pyBL.skin_friction": [[18, 3, 1, "", "c_f_Felsch"], [19, 3, 1, "", "c_f_LudwiegTillman"], [20, 3, 1, "", "c_f_White"]], "pyBL.stanford_olympics": [[22, 1, 1, "", "StanfordOlympics1968"]], "pyBL.stanford_olympics.StanfordOlympics1968": [[22, 2, 1, "", "G"], [22, 2, 1, "", "H_d"], [22, 2, 1, "", "H_k"], [22, 1, 1, "", "StationData"], [22, 2, 1, "", "U_e"], [22, 2, 1, "", "beta"], [22, 2, 1, "", "c_f"], [22, 2, 1, "", "c_f_E"], [22, 2, 1, "", "c_f_LT"], [22, 2, 1, "", "change_case_data"], [22, 2, 1, "", "dU_edx"], [22, 2, 1, "", "delta_d"], [22, 2, 1, "", "delta_k"], [22, 2, 1, "", "delta_m"], [22, 2, 1, "", "num_stations"], [22, 2, 1, "", "station"], [22, 2, 1, "", "velocity"], [22, 2, 1, "", "velocity_smooth"], [22, 2, 1, "", "x"]], "pyBL.thwaites_method": [[24, 1, 1, "", "ThwaitesMethod"], [25, 1, 1, "", "ThwaitesMethodLinear"], [26, 1, 1, "", "ThwaitesMethodNonlinear"]], "pyBL.thwaites_method.ThwaitesMethod": [[24, 2, 1, "", "D"], [24, 2, 1, "", "H_d"], [24, 2, 1, "", "H_k"], [24, 2, 1, "", "U_e"], [24, 2, 1, "", "V_e"], [24, 2, 1, "", "d2U_edx2"], [24, 2, 1, "", "dU_edx"], [24, 2, 1, "", "delta_d"], [24, 2, 1, "", "delta_k"], [24, 2, 1, "", "delta_m"], [24, 2, 1, "", "nu"], [24, 2, 1, "", "set_data_fits"], [24, 2, 1, "", "set_initial_condition"], [24, 2, 1, "", "set_initial_parameters"], [24, 2, 1, "", "set_velocity"], [24, 2, 1, "", "solve"], [24, 2, 1, "", "tau_w"]], "pyBL.thwaites_method.ThwaitesMethodLinear": [[25, 2, 1, "", "D"], [25, 2, 1, "", "H_d"], [25, 2, 1, "", "H_k"], [25, 2, 1, "", "U_e"], [25, 2, 1, "", "V_e"], [25, 2, 1, "", "d2U_edx2"], [25, 2, 1, "", "dU_edx"], [25, 2, 1, "", "delta_d"], [25, 2, 1, "", "delta_k"], [25, 2, 1, "", "delta_m"], [25, 2, 1, "", "nu"], [25, 2, 1, "", "set_data_fits"], [25, 2, 1, "", "set_initial_condition"], [25, 2, 1, "", "set_initial_parameters"], [25, 2, 1, "", "set_velocity"], [25, 2, 1, "", "solve"], [25, 2, 1, "", "tau_w"]], "pyBL.thwaites_method.ThwaitesMethodNonlinear": [[26, 2, 1, "", "D"], [26, 2, 1, "", "H_d"], [26, 2, 1, "", "H_k"], [26, 2, 1, "", "U_e"], [26, 2, 1, "", "V_e"], [26, 2, 1, "", "d2U_edx2"], [26, 2, 1, "", "dU_edx"], [26, 2, 1, "", "delta_d"], [26, 2, 1, "", "delta_k"], [26, 2, 1, "", "delta_m"], [26, 2, 1, "", "nu"], [26, 2, 1, "", "set_data_fits"], [26, 2, 1, "", "set_initial_condition"], [26, 2, 1, "", "set_initial_parameters"], [26, 2, 1, "", "set_velocity"], [26, 2, 1, "", "solve"], [26, 2, 1, "", "tau_w"]], "pyBL.xfoil_reader": [[28, 1, 1, "", "XFoilReader"]], "pyBL.xfoil_reader.XFoilReader": [[28, 1, 1, "", "AirfoilData"], [28, 2, 1, "", "H_d_lower"], [28, 2, 1, "", "H_d_upper"], [28, 2, 1, "", "H_d_wake"], [28, 2, 1, "", "H_k_lower"], [28, 2, 1, "", "H_k_upper"], [28, 2, 1, "", "K_lower"], [28, 2, 1, "", "K_upper"], [28, 2, 1, "", "P_lower"], [28, 2, 1, "", "P_upper"], [28, 2, 1, "", "U_e_lower"], [28, 2, 1, "", "U_e_upper"], [28, 2, 1, "", "U_e_wake"], [28, 1, 1, "", "WakeData"], [28, 2, 1, "", "c_f_lower"], [28, 2, 1, "", "c_f_upper"], [28, 2, 1, "", "change_case_data"], [28, 2, 1, "", "delta_d_lower"], [28, 2, 1, "", "delta_d_upper"], [28, 2, 1, "", "delta_d_wake"], [28, 2, 1, "", "delta_k_lower"], [28, 2, 1, "", "delta_k_upper"], [28, 2, 1, "", "delta_m_lower"], [28, 2, 1, "", "delta_m_upper"], [28, 2, 1, "", "delta_m_wake"], [28, 2, 1, "", "m_lower"], [28, 2, 1, "", "m_upper"], [28, 2, 1, "", "num_points_lower"], [28, 2, 1, "", "num_points_upper"], [28, 2, 1, "", "num_points_wake"], [28, 2, 1, "", "point_lower"], [28, 2, 1, "", "point_upper"], [28, 2, 1, "", "point_wake"], [28, 2, 1, "", "s_lower"], [28, 2, 1, "", "s_upper"], [28, 2, 1, "", "s_wake"], [28, 2, 1, "", "x_lower"], [28, 2, 1, "", "x_upper"], [28, 2, 1, "", "x_wake"], [28, 2, 1, "", "y_lower"], [28, 2, 1, "", "y_upper"], [28, 2, 1, "", "y_wake"]]}, "objtypes": {"0": "py:module", "1": "py:class", "2": "py:method", "3": "py:function"}, "objnames": {"0": ["py", "module", "Python module"], "1": ["py", "class", "Python class"], "2": ["py", "method", "Python method"], "3": ["py", "function", "Python function"]}, "titleterms": {"pybl": [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 34], "blasiu": [1, 2], "blasiussolut": 2, "dense_output_ppoli": 3, "falkner_skan": [4, 5], "falknerskansolut": 5, "head_method": [6, 7], "headmethod": 7, "ibl_method": [8, 9, 10, 11], "iblmethod": 9, "iblresult": 10, "ibltermev": 11, "initial_condit": [12, 13, 14, 15], "falknerskanstagnationcondit": 13, "initialcondit": 14, "manualcondit": 15, "skin_frict": [17, 18, 19, 20], "c_f_felsch": 18, "c_f_ludwiegtillman": 19, "c_f_white": 20, "stanford_olymp": [21, 22], "stanfordolympics1968": 22, "thwaites_method": [23, 24, 25, 26], "thwaitesmethod": 24, "thwaitesmethodlinear": 25, "thwaitesmethodnonlinear": 26, "xfoil_read": [27, 28], "xfoilread": 28, "api": 29, "contributor": [30, 34], "coven": 30, "code": [30, 31, 36], "conduct": [30, 31], "our": 30, "pledg": 30, "standard": 30, "respons": 30, "scope": 30, "enforc": 30, "attribut": 30, "contribut": 31, "style": 31, "guid": 31, "c": 31, "python": 31, "matlab": 31, "cmake": 31, "latex": 31, "file": 31, "script": 31, "other": [31, 36], "submit": 31, "chang": 31, "bug": 31, "report": 31, "test": 31, "develop": 32, "inform": [32, 35], "exampl": [33, 34], "summari": 34, "usag": 34, "version": [34, 36], "histori": 34, "licens": [34, 35, 36], "welcom": 34, "": 34, "document": 34, "content": 34, "indic": 34, "tabl": 34, "gnu": 36, "gener": 36, "public": 36, "preambl": 36, "term": 36, "AND": 36, "condit": 36, "0": 36, "definit": 36, "1": 36, "sourc": 36, "2": 36, "basic": 36, "permiss": 36, "3": 36, "protect": 36, "user": 36, "legal": 36, "right": 36, "from": 36, "anti": 36, "circumvent": 36, "law": 36, "4": 36, "convei": 36, "verbatim": 36, "copi": 36, "5": 36, "modifi": 36, "6": 36, "non": 36, "form": 36, "7": 36, "addit": 36, "8": 36, "termin": 36, "9": 36, "accept": 36, "Not": 36, "requir": 36, "have": 36, "10": 36, "automat": 36, "downstream": 36, "recipi": 36, "11": 36, "patent": 36, "12": 36, "No": 36, "surrend": 36, "freedom": 36, "13": 36, "us": 36, "affero": 36, "14": 36, "revis": 36, "thi": 36, "15": 36, "disclaim": 36, "warranti": 36, "16": 36, "limit": 36, "liabil": 36, "17": 36, "interpret": 36, "section": 36, "how": 36, "appli": 36, "These": 36, "your": 36, "new": 36, "program": 36}, "envversion": {"sphinx.domains.c": 2, "sphinx.domains.changeset": 1, "sphinx.domains.citation": 1, "sphinx.domains.cpp": 6, "sphinx.domains.index": 1, "sphinx.domains.javascript": 2, "sphinx.domains.math": 2, "sphinx.domains.python": 3, "sphinx.domains.rst": 2, "sphinx.domains.std": 2, "sphinx.ext.viewcode": 1, "sphinx": 56}})
+Search.setIndex({"alltitles": {"0. Definitions.": [[41, "definitions"]], "1. Source Code.": [[41, "source-code"]], "10. Automatic Licensing of Downstream Recipients.": [[41, "automatic-licensing-of-downstream-recipients"]], "11. Patents.": [[41, "patents"]], "12. No Surrender of Others\u2019 Freedom.": [[41, "no-surrender-of-others-freedom"]], "13. Use with the GNU Affero General Public License.": [[41, "use-with-the-gnu-affero-general-public-license"]], "14. Revised Versions of this License.": [[41, "revised-versions-of-this-license"]], "15. Disclaimer of Warranty.": [[41, "disclaimer-of-warranty"]], "16. Limitation of Liability.": [[41, "limitation-of-liability"]], "17. Interpretation of Sections 15 and 16.": [[41, "interpretation-of-sections-15-and-16"]], "2. Basic Permissions.": [[41, "basic-permissions"]], "3. Protecting Users\u2019 Legal Rights From Anti-Circumvention Law.": [[41, "protecting-users-legal-rights-from-anti-circumvention-law"]], "4. Conveying Verbatim Copies.": [[41, "conveying-verbatim-copies"]], "5. Conveying Modified Source Versions.": [[41, "conveying-modified-source-versions"]], "6. Conveying Non-Source Forms.": [[41, "conveying-non-source-forms"]], "7. Additional Terms.": [[41, "additional-terms"]], "8. Termination.": [[41, "termination"]], "9. Acceptance Not Required for Having Copies.": [[41, "acceptance-not-required-for-having-copies"]], "Attribution": [[35, "attribution"]], "Bug Reports": [[36, "bug-reports"]], "C/C++ Code": [[36, "c-c-code"]], "CMake Code": [[36, "cmake-code"]], "Code of Conduct": [[36, "code-of-conduct"]], "Contents:": [[39, null]], "Contributing": [[36, "contributing"]], "Contributor Covenant Code of Conduct": [[35, "contributor-covenant-code-of-conduct"]], "Contributors": [[39, "contributors"]], "Developer Information": [[37, "developer-information"]], "Enforcement": [[35, "enforcement"]], "Example Usage": [[39, "example-usage"]], "Examples": [[38, "examples"]], "GNU GENERAL PUBLIC LICENSE": [[41, "gnu-general-public-license"]], "How to Apply These Terms to Your New Programs": [[41, "how-to-apply-these-terms-to-your-new-programs"]], "Indices and tables": [[39, "indices-and-tables"]], "LaTeX Files": [[36, "latex-files"]], "License": [[39, "license"]], "License Information": [[40, "license-information"]], "Matlab Code": [[36, "matlab-code"]], "Our Pledge": [[35, "our-pledge"]], "Our Responsibilities": [[35, "our-responsibilities"]], "Our Standards": [[35, "our-standards"]], "Preamble": [[41, "preamble"]], "Python Code": [[36, "python-code"]], "Scope": [[35, "scope"]], "Scripts/Other Files": [[36, "scripts-other-files"]], "Style Guide": [[36, "style-guide"]], "Submitting Changes": [[36, "submitting-changes"]], "Summary": [[39, "summary"]], "TERMS AND CONDITIONS": [[41, "terms-and-conditions"]], "Testing": [[36, "testing"]], "Version History": [[39, "version-history"]], "Welcome to IBL\u2019s documentation!": [[39, "welcome-to-ibl-s-documentation"]], "ibl": [[0, "module-ibl"], [34, "ibl"]], "ibl.analytic": [[1, "module-ibl.analytic"]], "ibl.analytic.Analytic2dSimilarityIncompressible": [[2, "ibl-analytic-analytic2dsimilarityincompressible"]], "ibl.analytic.Blasius": [[3, "ibl-analytic-blasius"]], "ibl.analytic.FalknerSkan": [[4, "ibl-analytic-falknerskan"]], "ibl.head_method": [[5, "module-ibl.head_method"]], "ibl.head_method.HeadMethod": [[6, "ibl-head-method-headmethod"]], "ibl.ibl_method": [[7, "module-ibl.ibl_method"]], "ibl.ibl_method.IBLMethod": [[8, "ibl-ibl-method-iblmethod"]], "ibl.ibl_method.IntegrationResult": [[9, "ibl-ibl-method-integrationresult"]], "ibl.ibl_method.TermEvent": [[10, "ibl-ibl-method-termevent"]], "ibl.ibl_method.TermReason": [[11, "ibl-ibl-method-termreason"]], "ibl.initial_condition": [[12, "module-ibl.initial_condition"]], "ibl.initial_condition.FalknerSkanStagCondition": [[13, "ibl-initial-condition-falknerskanstagcondition"]], "ibl.initial_condition.InitialCondition": [[14, "ibl-initial-condition-initialcondition"]], "ibl.initial_condition.ManualCondition": [[15, "ibl-initial-condition-manualcondition"]], "ibl.reference": [[16, "module-ibl.reference"]], "ibl.reference.StanfordOlympics1968": [[17, "ibl-reference-stanfordolympics1968"]], "ibl.reference.StanfordOlympics1968SmoothVel": [[18, "ibl-reference-stanfordolympics1968smoothvel"]], "ibl.reference.StanfordOlympics1968StationData": [[19, "ibl-reference-stanfordolympics1968stationdata"]], "ibl.reference.XFoilAirfoilData": [[20, "ibl-reference-xfoilairfoildata"]], "ibl.reference.XFoilReader": [[21, "ibl-reference-xfoilreader"]], "ibl.reference.XFoilWakeData": [[22, "ibl-reference-xfoilwakedata"]], "ibl.skin_friction": [[23, "module-ibl.skin_friction"]], "ibl.skin_friction.felsch": [[24, "ibl-skin-friction-felsch"]], "ibl.skin_friction.ludwieg_tillman": [[25, "ibl-skin-friction-ludwieg-tillman"]], "ibl.skin_friction.white": [[26, "ibl-skin-friction-white"]], "ibl.thwaites_method": [[27, "module-ibl.thwaites_method"]], "ibl.thwaites_method.ThwaitesMethod": [[28, "ibl-thwaites-method-thwaitesmethod"]], "ibl.thwaites_method.ThwaitesMethodLinear": [[29, "ibl-thwaites-method-thwaitesmethodlinear"]], "ibl.thwaites_method.ThwaitesMethodNonlinear": [[30, "ibl-thwaites-method-thwaitesmethodnonlinear"]], "ibl.typing": [[31, "module-ibl.typing"]], "ibl.typing.InputParam": [[32, "ibl-typing-inputparam"]], "ibl.typing.SolutionFunc": [[33, "ibl-typing-solutionfunc"]]}, "docnames": ["api/_autosummary/ibl", "api/_autosummary/ibl.analytic", "api/_autosummary/ibl.analytic.Analytic2dSimilarityIncompressible", "api/_autosummary/ibl.analytic.Blasius", "api/_autosummary/ibl.analytic.FalknerSkan", "api/_autosummary/ibl.head_method", "api/_autosummary/ibl.head_method.HeadMethod", "api/_autosummary/ibl.ibl_method", "api/_autosummary/ibl.ibl_method.IBLMethod", "api/_autosummary/ibl.ibl_method.IntegrationResult", "api/_autosummary/ibl.ibl_method.TermEvent", "api/_autosummary/ibl.ibl_method.TermReason", "api/_autosummary/ibl.initial_condition", "api/_autosummary/ibl.initial_condition.FalknerSkanStagCondition", "api/_autosummary/ibl.initial_condition.InitialCondition", "api/_autosummary/ibl.initial_condition.ManualCondition", "api/_autosummary/ibl.reference", "api/_autosummary/ibl.reference.StanfordOlympics1968", "api/_autosummary/ibl.reference.StanfordOlympics1968SmoothVel", "api/_autosummary/ibl.reference.StanfordOlympics1968StationData", "api/_autosummary/ibl.reference.XFoilAirfoilData", "api/_autosummary/ibl.reference.XFoilReader", "api/_autosummary/ibl.reference.XFoilWakeData", "api/_autosummary/ibl.skin_friction", "api/_autosummary/ibl.skin_friction.felsch", "api/_autosummary/ibl.skin_friction.ludwieg_tillman", "api/_autosummary/ibl.skin_friction.white", "api/_autosummary/ibl.thwaites_method", "api/_autosummary/ibl.thwaites_method.ThwaitesMethod", "api/_autosummary/ibl.thwaites_method.ThwaitesMethodLinear", "api/_autosummary/ibl.thwaites_method.ThwaitesMethodNonlinear", "api/_autosummary/ibl.typing", "api/_autosummary/ibl.typing.InputParam", "api/_autosummary/ibl.typing.SolutionFunc", "api/modules", "developers/code_of_conduct", "developers/contributing", "developers/index", "examples/index", "index", "license/index", "license/license"], "envversion": {"sphinx": 61, "sphinx.domains.c": 3, "sphinx.domains.changeset": 1, "sphinx.domains.citation": 1, "sphinx.domains.cpp": 9, "sphinx.domains.index": 1, "sphinx.domains.javascript": 3, "sphinx.domains.math": 2, "sphinx.domains.python": 4, "sphinx.domains.rst": 2, "sphinx.domains.std": 2}, "filenames": ["api/_autosummary/ibl.rst", "api/_autosummary/ibl.analytic.rst", "api/_autosummary/ibl.analytic.Analytic2dSimilarityIncompressible.rst", "api/_autosummary/ibl.analytic.Blasius.rst", "api/_autosummary/ibl.analytic.FalknerSkan.rst", "api/_autosummary/ibl.head_method.rst", "api/_autosummary/ibl.head_method.HeadMethod.rst", "api/_autosummary/ibl.ibl_method.rst", "api/_autosummary/ibl.ibl_method.IBLMethod.rst", "api/_autosummary/ibl.ibl_method.IntegrationResult.rst", "api/_autosummary/ibl.ibl_method.TermEvent.rst", "api/_autosummary/ibl.ibl_method.TermReason.rst", "api/_autosummary/ibl.initial_condition.rst", "api/_autosummary/ibl.initial_condition.FalknerSkanStagCondition.rst", "api/_autosummary/ibl.initial_condition.InitialCondition.rst", "api/_autosummary/ibl.initial_condition.ManualCondition.rst", "api/_autosummary/ibl.reference.rst", "api/_autosummary/ibl.reference.StanfordOlympics1968.rst", "api/_autosummary/ibl.reference.StanfordOlympics1968SmoothVel.rst", "api/_autosummary/ibl.reference.StanfordOlympics1968StationData.rst", "api/_autosummary/ibl.reference.XFoilAirfoilData.rst", "api/_autosummary/ibl.reference.XFoilReader.rst", "api/_autosummary/ibl.reference.XFoilWakeData.rst", "api/_autosummary/ibl.skin_friction.rst", "api/_autosummary/ibl.skin_friction.felsch.rst", "api/_autosummary/ibl.skin_friction.ludwieg_tillman.rst", "api/_autosummary/ibl.skin_friction.white.rst", "api/_autosummary/ibl.thwaites_method.rst", "api/_autosummary/ibl.thwaites_method.ThwaitesMethod.rst", "api/_autosummary/ibl.thwaites_method.ThwaitesMethodLinear.rst", "api/_autosummary/ibl.thwaites_method.ThwaitesMethodNonlinear.rst", "api/_autosummary/ibl.typing.rst", "api/_autosummary/ibl.typing.InputParam.rst", "api/_autosummary/ibl.typing.SolutionFunc.rst", "api/modules.rst", "developers/code_of_conduct.rst", "developers/contributing.rst", "developers/index.rst", "examples/index.rst", "index.rst", "license/index.rst", "license/license.rst"], "indexentries": {}, "objects": {"": [[0, 0, 0, "-", "ibl"]], "ibl": [[1, 0, 0, "-", "analytic"], [5, 0, 0, "-", "head_method"], [7, 0, 0, "-", "ibl_method"], [12, 0, 0, "-", "initial_condition"], [16, 0, 0, "-", "reference"], [23, 0, 0, "-", "skin_friction"], [27, 0, 0, "-", "thwaites_method"], [31, 0, 0, "-", "typing"]], "ibl.analytic": [[2, 1, 1, "", "Analytic2dSimilarityIncompressible"], [3, 1, 1, "", "Blasius"], [4, 1, 1, "", "FalknerSkan"]], "ibl.analytic.Analytic2dSimilarityIncompressible": [[2, 2, 1, "", "__init__"], [2, 2, 1, "", "delta_d"], [2, 2, 1, "", "delta_k"], [2, 2, 1, "", "delta_m"], [2, 2, 1, "", "delta_s"], [2, 2, 1, "", "dissipation"], [2, 2, 1, "", "eta"], [2, 3, 1, "", "eta_d"], [2, 3, 1, "", "eta_inf"], [2, 3, 1, "", "eta_k"], [2, 3, 1, "", "eta_m"], [2, 3, 1, "", "eta_s"], [2, 2, 1, "", "f"], [2, 2, 1, "", "f_p"], [2, 2, 1, "", "f_pp"], [2, 3, 1, "", "fw_pp"], [2, 3, 1, "", "nu_ref"], [2, 2, 1, "", "set_solution_parameters"], [2, 2, 1, "", "shape_d"], [2, 2, 1, "", "shape_k"], [2, 2, 1, "", "tau_w"], [2, 2, 1, "", "u"], [2, 2, 1, "", "u_e"], [2, 3, 1, "", "u_ref"], [2, 2, 1, "", "v"], [2, 2, 1, "", "v_e"]], "ibl.analytic.Blasius": [[3, 2, 1, "", "__init__"], [3, 2, 1, "", "delta_d"], [3, 2, 1, "", "delta_k"], [3, 2, 1, "", "delta_m"], [3, 2, 1, "", "delta_s"], [3, 2, 1, "", "dissipation"], [3, 2, 1, "", "eta"], [3, 3, 1, "", "eta_d"], [3, 3, 1, "", "eta_inf"], [3, 3, 1, "", "eta_inf_default"], [3, 3, 1, "", "eta_k"], [3, 3, 1, "", "eta_m"], [3, 3, 1, "", "eta_s"], [3, 2, 1, "", "f"], [3, 2, 1, "", "f_p"], [3, 2, 1, "", "f_pp"], [3, 3, 1, "", "fw_pp"], [3, 3, 1, "", "fw_pp_default"], [3, 3, 1, "", "nu_ref"], [3, 2, 1, "", "set_solution_parameters"], [3, 2, 1, "", "shape_d"], [3, 2, 1, "", "shape_k"], [3, 2, 1, "", "tau_w"], [3, 2, 1, "", "u"], [3, 2, 1, "", "u_e"], [3, 3, 1, "", "u_ref"], [3, 2, 1, "", "v"], [3, 2, 1, "", "v_e"]], "ibl.analytic.FalknerSkan": [[4, 2, 1, "", "__init__"], [4, 3, 1, "", "beta"], [4, 2, 1, "", "delta_d"], [4, 2, 1, "", "delta_k"], [4, 2, 1, "", "delta_m"], [4, 2, 1, "", "delta_s"], [4, 2, 1, "", "dissipation"], [4, 2, 1, "", "eta"], [4, 3, 1, "", "eta_d"], [4, 3, 1, "", "eta_inf"], [4, 3, 1, "", "eta_k"], [4, 3, 1, "", "eta_m"], [4, 3, 1, "", "eta_s"], [4, 2, 1, "", "f"], [4, 2, 1, "", "f_p"], [4, 2, 1, "", "f_pp"], [4, 3, 1, "", "fw_pp"], [4, 3, 1, "", "m"], [4, 3, 1, "", "nu_ref"], [4, 2, 1, "", "reset_beta"], [4, 2, 1, "", "reset_m"], [4, 2, 1, "", "set_solution_parameters"], [4, 2, 1, "", "shape_d"], [4, 2, 1, "", "shape_k"], [4, 2, 1, "", "tau_w"], [4, 2, 1, "", "u"], [4, 2, 1, "", "u_e"], [4, 3, 1, "", "u_ref"], [4, 2, 1, "", "v"], [4, 2, 1, "", "v_e"]], "ibl.head_method": [[6, 1, 1, "", "HeadMethod"]], "ibl.head_method.HeadMethod": [[6, 2, 1, "", "__init__"], [6, 2, 1, "", "d2u_e"], [6, 2, 1, "", "delta_d"], [6, 2, 1, "", "delta_k"], [6, 2, 1, "", "delta_m"], [6, 2, 1, "", "dissipation"], [6, 2, 1, "", "du_e"], [6, 3, 1, "", "initial_delta_m"], [6, 3, 1, "", "initial_shape_d"], [6, 3, 1, "", "nu"], [6, 2, 1, "", "set_initial_condition"], [6, 2, 1, "", "set_shape_d_critical"], [6, 2, 1, "", "set_velocity"], [6, 2, 1, "", "shape_d"], [6, 2, 1, "", "shape_k"], [6, 2, 1, "", "solve"], [6, 2, 1, "", "tau_w"], [6, 2, 1, "", "u_e"], [6, 2, 1, "", "v_e"]], "ibl.ibl_method": [[8, 1, 1, "", "IBLMethod"], [9, 1, 1, "", "IntegrationResult"], [10, 1, 1, "", "TermEvent"], [11, 1, 1, "", "TermReason"]], "ibl.ibl_method.IBLMethod": [[8, 2, 1, "", "__init__"], [8, 2, 1, "", "d2u_e"], [8, 2, 1, "", "delta_d"], [8, 2, 1, "", "delta_k"], [8, 2, 1, "", "delta_m"], [8, 2, 1, "", "dissipation"], [8, 2, 1, "", "du_e"], [8, 3, 1, "", "nu"], [8, 2, 1, "", "set_initial_condition"], [8, 2, 1, "", "set_velocity"], [8, 2, 1, "", "shape_d"], [8, 2, 1, "", "shape_k"], [8, 2, 1, "", "solve"], [8, 2, 1, "", "tau_w"], [8, 2, 1, "", "u_e"], [8, 2, 1, "", "v_e"]], "ibl.ibl_method.IntegrationResult": [[9, 2, 1, "", "__init__"], [9, 3, 1, "", "f_end"], [9, 3, 1, "", "message"], [9, 3, 1, "", "status"], [9, 3, 1, "", "success"], [9, 3, 1, "", "x_end"]], "ibl.ibl_method.TermEvent": [[10, 2, 1, "", "__init__"], [10, 2, 1, "", "event_info"], [10, 4, 1, "", "terminal"]], "ibl.ibl_method.TermReason": [[11, 4, 1, "", "REACHED_END"], [11, 4, 1, "", "SEPARATED"], [11, 4, 1, "", "TRANSITIONED"], [11, 4, 1, "", "UNKNOWN"], [11, 2, 1, "", "__init__"], [11, 2, 1, "", "as_integer_ratio"], [11, 2, 1, "", "bit_length"], [11, 2, 1, "", "conjugate"], [11, 4, 1, "", "denominator"], [11, 2, 1, "", "from_bytes"], [11, 4, 1, "", "imag"], [11, 4, 1, "", "numerator"], [11, 4, 1, "", "real"], [11, 2, 1, "", "to_bytes"]], "ibl.initial_condition": [[13, 1, 1, "", "FalknerSkanStagCondition"], [14, 1, 1, "", "InitialCondition"], [15, 1, 1, "", "ManualCondition"]], "ibl.initial_condition.FalknerSkanStagCondition": [[13, 2, 1, "", "__init__"], [13, 2, 1, "", "delta_d"], [13, 2, 1, "", "delta_k"], [13, 2, 1, "", "delta_m"], [13, 3, 1, "", "du_e"], [13, 3, 1, "", "nu"], [13, 2, 1, "", "shape_d"], [13, 2, 1, "", "shape_k"]], "ibl.initial_condition.InitialCondition": [[14, 2, 1, "", "__init__"], [14, 2, 1, "", "delta_d"], [14, 2, 1, "", "delta_k"], [14, 2, 1, "", "delta_m"], [14, 3, 1, "", "du_e"], [14, 3, 1, "", "nu"], [14, 2, 1, "", "shape_d"], [14, 2, 1, "", "shape_k"]], "ibl.initial_condition.ManualCondition": [[15, 2, 1, "", "__init__"], [15, 4, 1, "", "del_d"], [15, 4, 1, "", "del_k"], [15, 4, 1, "", "del_m"], [15, 2, 1, "", "delta_d"], [15, 2, 1, "", "delta_k"], [15, 2, 1, "", "delta_m"], [15, 3, 1, "", "du_e"], [15, 3, 1, "", "nu"], [15, 2, 1, "", "shape_d"], [15, 2, 1, "", "shape_k"]], "ibl.reference": [[17, 1, 1, "", "StanfordOlympics1968"], [18, 1, 1, "", "StanfordOlympics1968SmoothVel"], [19, 1, 1, "", "StanfordOlympics1968StationData"], [20, 1, 1, "", "XFoilAirfoilData"], [21, 1, 1, "", "XFoilReader"], [22, 1, 1, "", "XFoilWakeData"]], "ibl.reference.StanfordOlympics1968": [[17, 2, 1, "", "__init__"], [17, 2, 1, "", "beta_eq"], [17, 2, 1, "", "c_f"], [17, 2, 1, "", "c_f_exp"], [17, 2, 1, "", "c_f_lt"], [17, 3, 1, "", "case"], [17, 2, 1, "", "delta_c"], [17, 2, 1, "", "delta_d"], [17, 2, 1, "", "delta_k"], [17, 2, 1, "", "delta_m"], [17, 2, 1, "", "du_e"], [17, 2, 1, "", "du_e_smooth"], [17, 2, 1, "", "nu"], [17, 3, 1, "", "nu_ref"], [17, 2, 1, "", "re_delta_d"], [17, 2, 1, "", "re_delta_m"], [17, 2, 1, "", "shape_d"], [17, 2, 1, "", "shape_eq"], [17, 2, 1, "", "shape_k"], [17, 2, 1, "", "station"], [17, 2, 1, "", "station_count"], [17, 2, 1, "", "u"], [17, 2, 1, "", "u_defect"], [17, 2, 1, "", "u_e"], [17, 2, 1, "", "u_e_smooth"], [17, 2, 1, "", "u_plus"], [17, 2, 1, "", "u_star"], [17, 2, 1, "", "x"], [17, 2, 1, "", "x_smooth"], [17, 2, 1, "", "y"], [17, 2, 1, "", "y_c"], [17, 2, 1, "", "y_plus"]], "ibl.reference.StanfordOlympics1968SmoothVel": [[18, 2, 1, "", "__init__"], [18, 3, 1, "", "du_e"], [18, 2, 1, "", "reset"], [18, 2, 1, "", "size"], [18, 3, 1, "", "u_e"], [18, 3, 1, "", "x"]], "ibl.reference.StanfordOlympics1968StationData": [[19, 2, 1, "", "__init__"], [19, 3, 1, "", "beta_eq"], [19, 3, 1, "", "c_f"], [19, 3, 1, "", "c_f_exp"], [19, 3, 1, "", "c_f_lt"], [19, 3, 1, "", "delta_c"], [19, 3, 1, "", "delta_d"], [19, 3, 1, "", "delta_k"], [19, 3, 1, "", "delta_m"], [19, 3, 1, "", "du_e"], [19, 3, 1, "", "nu"], [19, 3, 1, "", "re_delta_d"], [19, 3, 1, "", "re_delta_m"], [19, 2, 1, "", "reset"], [19, 2, 1, "", "sample_count"], [19, 3, 1, "", "shape_d"], [19, 3, 1, "", "shape_eq"], [19, 3, 1, "", "shape_k"], [19, 3, 1, "", "u"], [19, 3, 1, "", "u_defect"], [19, 3, 1, "", "u_e"], [19, 3, 1, "", "u_plus"], [19, 3, 1, "", "u_star"], [19, 3, 1, "", "x"], [19, 3, 1, "", "y"], [19, 3, 1, "", "y_c"], [19, 3, 1, "", "y_plus"]], "ibl.reference.XFoilAirfoilData": [[20, 2, 1, "", "__init__"], [20, 3, 1, "", "c_f"], [20, 3, 1, "", "delta_d"], [20, 3, 1, "", "delta_m"], [20, 3, 1, "", "ke_defect"], [20, 3, 1, "", "mass_defect"], [20, 3, 1, "", "mom_defect"], [20, 2, 1, "", "reset"], [20, 3, 1, "", "s"], [20, 3, 1, "", "shape_d"], [20, 3, 1, "", "shape_k"], [20, 3, 1, "", "u_e_rel"], [20, 3, 1, "", "x"], [20, 3, 1, "", "y"]], "ibl.reference.XFoilReader": [[21, 2, 1, "", "__init__"], [21, 3, 1, "", "alpha"], [21, 3, 1, "", "c"], [21, 2, 1, "", "c_f_lower"], [21, 2, 1, "", "c_f_upper"], [21, 2, 1, "", "delta_d_lower"], [21, 2, 1, "", "delta_d_upper"], [21, 2, 1, "", "delta_d_wake"], [21, 2, 1, "", "delta_k_lower"], [21, 2, 1, "", "delta_k_upper"], [21, 2, 1, "", "delta_m_lower"], [21, 2, 1, "", "delta_m_upper"], [21, 2, 1, "", "delta_m_wake"], [21, 3, 1, "", "filename"], [21, 2, 1, "", "ke_defect_lower"], [21, 2, 1, "", "ke_defect_upper"], [21, 2, 1, "", "lower"], [21, 2, 1, "", "lower_count"], [21, 2, 1, "", "mass_defect_lower"], [21, 2, 1, "", "mass_defect_upper"], [21, 2, 1, "", "mom_defect_lower"], [21, 2, 1, "", "mom_defect_upper"], [21, 3, 1, "", "n_trans"], [21, 3, 1, "", "name"], [21, 3, 1, "", "reynolds"], [21, 2, 1, "", "s_lower"], [21, 2, 1, "", "s_upper"], [21, 2, 1, "", "s_wake"], [21, 2, 1, "", "shape_d_lower"], [21, 2, 1, "", "shape_d_upper"], [21, 2, 1, "", "shape_d_wake"], [21, 2, 1, "", "shape_k_lower"], [21, 2, 1, "", "shape_k_upper"], [21, 2, 1, "", "u_e_lower"], [21, 2, 1, "", "u_e_upper"], [21, 2, 1, "", "u_e_wake"], [21, 3, 1, "", "u_ref"], [21, 2, 1, "", "upper"], [21, 2, 1, "", "upper_count"], [21, 2, 1, "", "wake"], [21, 2, 1, "", "wake_count"], [21, 2, 1, "", "x_lower"], [21, 3, 1, "", "x_trans_lower"], [21, 3, 1, "", "x_trans_upper"], [21, 2, 1, "", "x_upper"], [21, 2, 1, "", "x_wake"], [21, 2, 1, "", "y_lower"], [21, 2, 1, "", "y_upper"], [21, 2, 1, "", "y_wake"]], "ibl.reference.XFoilWakeData": [[22, 2, 1, "", "__init__"], [22, 3, 1, "", "delta_d"], [22, 3, 1, "", "delta_m"], [22, 2, 1, "", "reset"], [22, 3, 1, "", "s"], [22, 3, 1, "", "shape_d"], [22, 3, 1, "", "u_e_rel"], [22, 3, 1, "", "x"], [22, 3, 1, "", "y"]], "ibl.skin_friction": [[24, 5, 1, "", "felsch"], [25, 5, 1, "", "ludwieg_tillman"], [26, 5, 1, "", "white"]], "ibl.thwaites_method": [[28, 1, 1, "", "ThwaitesMethod"], [29, 1, 1, "", "ThwaitesMethodLinear"], [30, 1, 1, "", "ThwaitesMethodNonlinear"]], "ibl.thwaites_method.ThwaitesMethod": [[28, 2, 1, "", "__init__"], [28, 2, 1, "", "d2u_e"], [28, 2, 1, "", "delta_d"], [28, 2, 1, "", "delta_k"], [28, 2, 1, "", "delta_m"], [28, 2, 1, "", "dissipation"], [28, 2, 1, "", "du_e"], [28, 3, 1, "", "initial_delta_m"], [28, 3, 1, "", "nu"], [28, 2, 1, "", "set_data_fits"], [28, 2, 1, "", "set_initial_condition"], [28, 2, 1, "", "set_velocity"], [28, 2, 1, "", "shape_d"], [28, 2, 1, "", "shape_k"], [28, 2, 1, "", "solve"], [28, 2, 1, "", "tau_w"], [28, 2, 1, "", "u_e"], [28, 2, 1, "", "v_e"]], "ibl.thwaites_method.ThwaitesMethodLinear": [[29, 2, 1, "", "__init__"], [29, 2, 1, "", "d2u_e"], [29, 2, 1, "", "delta_d"], [29, 2, 1, "", "delta_k"], [29, 2, 1, "", "delta_m"], [29, 2, 1, "", "dissipation"], [29, 2, 1, "", "du_e"], [29, 3, 1, "", "initial_delta_m"], [29, 3, 1, "", "nu"], [29, 2, 1, "", "set_data_fits"], [29, 2, 1, "", "set_initial_condition"], [29, 2, 1, "", "set_velocity"], [29, 2, 1, "", "shape_d"], [29, 2, 1, "", "shape_k"], [29, 2, 1, "", "solve"], [29, 2, 1, "", "tau_w"], [29, 2, 1, "", "u_e"], [29, 2, 1, "", "v_e"]], "ibl.thwaites_method.ThwaitesMethodNonlinear": [[30, 2, 1, "", "__init__"], [30, 2, 1, "", "d2u_e"], [30, 2, 1, "", "delta_d"], [30, 2, 1, "", "delta_k"], [30, 2, 1, "", "delta_m"], [30, 2, 1, "", "dissipation"], [30, 2, 1, "", "du_e"], [30, 3, 1, "", "initial_delta_m"], [30, 3, 1, "", "nu"], [30, 2, 1, "", "set_data_fits"], [30, 2, 1, "", "set_initial_condition"], [30, 2, 1, "", "set_velocity"], [30, 2, 1, "", "shape_d"], [30, 2, 1, "", "shape_k"], [30, 2, 1, "", "solve"], [30, 2, 1, "", "tau_w"], [30, 2, 1, "", "u_e"], [30, 2, 1, "", "v_e"]], "ibl.typing": [[32, 6, 1, "", "InputParam"], [33, 6, 1, "", "SolutionFunc"]]}, "objnames": {"0": ["py", "module", "Python module"], "1": ["py", "class", "Python class"], "2": ["py", "method", "Python method"], "3": ["py", "property", "Python property"], "4": ["py", "attribute", "Python attribute"], "5": ["py", "function", "Python function"], "6": ["py", "data", "Python data"]}, "objtypes": {"0": "py:module", "1": "py:class", "2": "py:method", "3": "py:property", "4": "py:attribute", "5": "py:function", "6": "py:data"}, "terms": {"": [5, 6, 8, 9, 11, 16, 17, 18, 19, 20, 21, 22, 28, 29, 30, 35, 41], "0": [3, 4, 6, 11, 28, 29, 30, 39], "0904295": 4, "0b100101": 11, "1": [6, 11, 17, 18, 19, 28, 29, 30, 35, 39], "10": 11, "1949": [28, 30, 39], "1950": 25, "1958": [6, 39], "1968": [16, 17, 18, 19, 24], "1977": [28, 29, 30], "19884": 4, "1991": 28, "1996": 41, "1cd": 19, "2": [4, 6, 8, 11, 17, 19, 28, 29, 30, 39], "20": 41, "2005": 28, "2007": 41, "2011": [26, 28, 29, 30], "2014": 28, "2022": [28, 29, 30], "245": [28, 39], "28": 41, "280": [28, 39], "29": 41, "2d": 2, "3": [11, 28, 29, 30, 39, 40], "30": 41, "3152": 39, "37": 11, "4": [6, 35, 39], "45": 29, "5": 39, "6": [11, 29, 39], "60": 41, "6b": 41, "6d": 41, "8": 36, "A": [36, 39, 40, 41], "AND": [39, 40], "AS": 41, "And": 41, "At": 8, "BE": 41, "BEING": 41, "BUT": 41, "BY": 41, "Being": 35, "But": 41, "By": [8, 41], "FOR": [39, 40, 41], "For": [10, 35, 36, 39, 41], "IF": 41, "IN": 41, "If": [2, 3, 4, 8, 11, 18, 19, 20, 21, 22, 24, 25, 26, 36, 39, 41], "In": [8, 28, 35, 36, 39, 41], "It": [4, 28, 39, 41], "NO": 41, "NOT": 41, "No": [28, 39], "Not": 9, "OF": 41, "OR": 41, "Of": 41, "One": 39, "SUCH": 41, "Such": 41, "THE": 41, "THERE": 41, "TO": 41, "The": [2, 3, 4, 6, 8, 9, 10, 11, 16, 17, 21, 28, 29, 30, 35, 36, 39, 41], "There": [6, 7, 8, 27, 28, 29, 30], "These": [8, 12, 17, 36, 39, 40], "To": [9, 11, 36, 41], "WILL": 41, "WITH": 41, "__init__": [2, 3, 4, 6, 8, 9, 10, 11, 13, 14, 15, 17, 18, 19, 20, 21, 22, 28, 29, 30], "_call_impl": 10, "abc": [2, 8, 10, 14], "abil": 41, "about": [9, 10, 35, 41], "abov": 41, "absenc": 41, "absolut": 41, "abstract": [8, 10, 12, 14], "abus": [35, 41], "accept": 35, "access": 41, "accompani": 41, "accord": 41, "account": 35, "accur": [3, 8, 30], "achiev": 41, "acknowledg": 41, "acquir": 41, "across": 41, "act": 35, "action": [35, 41], "activ": 41, "actual": 41, "ad": [39, 41], "adapt": [35, 41], "add": 41, "addit": [28, 36, 39], "address": [35, 41], "adopt": 41, "advanc": 35, "advers": 41, "advis": 41, "aeronaut": [28, 39], "affect": 41, "affirm": 41, "afosr": [16, 17], "after": [9, 41], "ag": [35, 36], "against": 41, "aggreg": 41, "agre": 41, "agreement": 41, "aim": 41, "airfoil": [20, 21], "airfoildata": 21, "airofil": 21, "algorithm": 28, "alia": [32, 33], "align": 35, "all": [7, 8, 16, 17, 19, 27, 35, 41], "alleg": 41, "allow": [8, 41], "along": [17, 20, 22, 28, 39, 41], "alpha": [2, 21], "alreadi": 41, "also": [17, 35, 36, 41], "altern": 41, "although": 41, "among": 41, "amplif": 21, "an": [0, 3, 7, 8, 10, 11, 17, 21, 28, 29, 30, 35, 36, 39, 41], "analysi": [6, 8, 28, 29, 30], "analyt": 39, "analytic2dsimilarityincompress": [3, 4], "analyz": [21, 36], "ancillari": 41, "angl": [4, 21], "ani": [2, 3, 4, 6, 8, 9, 10, 11, 18, 19, 28, 29, 30, 32, 35, 39, 40, 41], "anoth": 6, "answer": 35, "antideriv": 8, "anyon": 41, "anyth": [37, 41], "api": 39, "appear": [35, 36], "appli": [35, 39, 40], "applic": 41, "appoint": 35, "appropri": [4, 35, 41], "approxim": [4, 8, 28, 29, 30, 39, 41], "ar": [2, 3, 4, 6, 7, 8, 9, 10, 11, 16, 17, 21, 24, 25, 26, 27, 28, 29, 30, 35, 36, 37, 39, 41], "arc": [21, 39], "arclength": [20, 22], "area": 41, "argument": 11, "aris": 41, "arrai": [6, 8, 11, 24, 25, 26, 28, 29, 30, 32], "arrang": 41, "articl": 41, "as_integer_ratio": 11, "ask": 41, "assert": 41, "asset": 41, "associ": [2, 3, 4, 6, 7, 8, 10, 19, 28, 29, 30, 41], "assum": [8, 21, 41], "assumpt": [27, 41], "assur": 41, "attach": [39, 41], "attack": [21, 35], "attempt": 41, "attent": 35, "attribut": [2, 3, 4, 6, 8, 9, 11, 13, 14, 15, 17, 18, 19, 20, 21, 22, 28, 29, 30, 31, 41], "author": 41, "avail": [35, 36, 41], "aviat": 39, "avoid": 41, "awai": 41, "b": [39, 41], "back": [2, 3, 4], "ban": 35, "base": [2, 3, 4, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 17, 18, 19, 20, 21, 22, 24, 25, 26, 27, 28, 29, 30, 39, 41], "becaus": [6, 8, 11, 28, 29, 30, 41], "been": [6, 8, 28, 29, 30, 41], "befor": [6, 8, 28, 29, 30, 36], "begin": 11, "behalf": 41, "behavior": [16, 35], "being": [2, 3, 4, 8, 21, 41], "believ": 41, "below": 41, "benefit": 41, "best": [35, 36, 41], "beta": [2, 4, 19], "beta_eq": [17, 19], "better": 27, "between": [8, 17, 39, 41], "beyond": 41, "big": 11, "bin": 11, "binari": 11, "bit": 11, "bit_length": 11, "bodi": [35, 36, 39, 41], "bondari": 6, "bool": [9, 10, 18, 19], "both": [2, 3, 4, 8, 17, 24, 25, 26, 36, 41], "boundari": [0, 1, 2, 3, 4, 5, 6, 7, 8, 11, 12, 14, 16, 17, 19, 21, 23, 24, 25, 26, 27, 28, 29, 30, 39], "box": 41, "bradshaw": [28, 29, 30], "brief": 41, "buffer": 11, "built": 11, "bunch": 9, "busi": 41, "byte": 11, "bytearrai": 11, "byteord": 11, "c": [21, 41], "c_f": [17, 19, 20], "c_f_exp": [17, 19], "c_f_lower": 21, "c_f_lt": [17, 19], "c_f_upper": 21, "calcul": [2, 3, 4, 6, 8, 23, 24, 25, 26, 28, 29, 30, 39], "call": [8, 10, 41], "callabl": [8, 28, 29, 30, 33], "can": [2, 3, 4, 6, 8, 10, 12, 16, 28, 29, 30, 32, 36, 39, 40, 41], "cannot": 41, "capabl": 37, "captur": 30, "carri": 41, "cartesian": [2, 3, 4], "case": [8, 16, 17, 19, 21, 27, 32, 41], "caus": [2, 3, 4, 10, 41], "cd": 19, "ceas": 41, "cebeci": [28, 29, 30], "certain": 41, "cessat": 41, "cf": [19, 20], "cfe": 19, "cflt": 19, "chang": [8, 10, 13, 14, 15, 17, 18, 19, 39, 41], "character": [16, 41], "characterist": [35, 36], "characterst": 7, "charg": 41, "check": [36, 39], "child": [6, 8, 28, 29, 30], "choos": 41, "chord": [20, 21, 22], "circumst": [2, 3, 4, 35, 41], "civil": 41, "cl": 19, "claim": 41, "clarifi": 35, "class": [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 27, 28, 29, 30, 39, 41], "clauser": [17, 19], "clean": 39, "clear": 41, "clearli": 41, "close": 41, "code": [7, 16, 37, 39], "codit": 39, "coeffici": [17, 19, 20, 21, 23, 24, 25, 26], "collect": [16, 41], "combin": 41, "come": [17, 41], "command": 41, "comment": 35, "commerci": 41, "commit": [35, 36, 41], "common": [7, 8, 12, 35, 36, 41], "commun": [35, 36, 41], "comparison": 16, "compil": [17, 41], "complaint": 35, "complement": 11, "complet": [6, 7, 8, 9, 11, 28, 29, 30], "complex": 11, "compli": 41, "complianc": 41, "compon": [2, 3, 4, 41], "comput": [16, 17, 41], "concern": 41, "concret": [27, 39], "condit": [2, 3, 4, 6, 8, 12, 13, 14, 15, 28, 29, 30, 39, 40], "condtion": [13, 14, 15], "conduct": [37, 39], "confer": [16, 17], "confidenti": 35, "configur": [6, 8, 28, 29, 30, 39], "conjug": 11, "connect": 41, "consequ": 41, "consequenti": 41, "consid": [35, 41], "consist": [17, 41], "conspicu": 41, "constantli": 41, "constitut": 41, "constru": 41, "construct": [35, 39], "consum": 41, "contact": [35, 41], "contain": [5, 16, 19, 20, 21, 22, 27, 36, 41], "content": 41, "context": 41, "continu": 41, "contractu": 41, "contradict": 41, "contrast": 41, "contribut": [35, 37, 39], "contributor": [36, 37, 41], "control": 41, "conveni": 41, "convert": 11, "convey": 41, "coordin": [2, 3, 4, 18, 19, 20, 22], "copi": 39, "copyleft": 41, "copyright": 41, "correct": [35, 41], "correspond": [2, 3, 4, 17, 24, 25, 26, 39, 41], "cost": 41, "could": [35, 39, 41], "council": 39, "counterclaim": 41, "countri": 41, "coupl": 16, "cours": 41, "court": 41, "coven": [37, 39, 41], "cover": 41, "coverag": 41, "creat": [8, 12, 35], "criteria": 39, "criterion": 41, "critic": [6, 35], "cross": 41, "cubic": 8, "cure": 41, "currat": 16, "current": [16, 39], "curv": [18, 28, 29, 30], "custom": [12, 41], "customarili": 41, "d": [29, 30, 39], "d2u_e": [6, 8, 28, 29, 30], "d2u_edx2": [6, 28, 29, 30], "dai": 41, "damag": 41, "danger": 41, "data": [5, 7, 12, 16, 17, 18, 19, 20, 21, 22, 24, 25, 26, 27, 28, 29, 30, 39, 41], "data_fit": [28, 29, 30], "date": 41, "david": 39, "decemb": 41, "decid": 41, "declin": 41, "dedic": 16, "deem": [35, 41], "default": [2, 3, 4, 6, 8, 9, 17, 21, 28, 29, 30], "defect": [17, 19, 20, 21, 41], "defens": 41, "defin": [2, 3, 35, 41], "definit": [6, 8, 17, 18, 19, 28, 29, 30], "degre": 8, "del": 19, "del_d": 15, "del_k": 15, "del_m": 15, "delta_": [2, 3, 4], "delta_c": [17, 19], "delta_d": [2, 3, 4, 6, 8, 13, 14, 15, 17, 19, 20, 22, 28, 29, 30], "delta_d_low": 21, "delta_d_upp": 21, "delta_d_wak": 21, "delta_k": [2, 3, 4, 6, 8, 13, 14, 15, 17, 19, 28, 29, 30], "delta_k_low": 21, "delta_k_upp": 21, "delta_m": [2, 3, 4, 6, 8, 13, 14, 15, 17, 19, 20, 22, 28, 29, 30], "delta_m0": 39, "delta_m_low": 21, "delta_m_upp": 21, "delta_m_wak": 21, "deni": 41, "denomin": [11, 41], "dens": [2, 3, 4, 8], "densiti": [2, 3, 4, 6, 8, 28, 29, 30, 39], "depend": 8, "depriv": 41, "deriv": [2, 3, 4, 6, 8, 10, 18, 28, 29, 30], "derogatori": 35, "descript": 9, "design": 41, "desir": [6, 8, 17, 28, 29, 30], "detail": [6, 8, 9, 10, 20, 28, 29, 30, 35, 39, 40, 41], "determin": [6, 11, 14, 35, 41], "develop": [39, 41], "devic": 41, "differ": [6, 8, 28, 29, 30, 35, 41], "differenti": [8, 29, 30], "digit": 17, "dimension": [2, 3, 4, 17, 19], "direct": [2, 3, 4, 8, 41], "directli": [39, 41], "directori": 36, "disabl": [35, 36], "discriminatori": 41, "discuss": 36, "dispac": 6, "displac": [2, 3, 4, 6, 8, 13, 14, 15, 17, 19, 20, 21, 22, 23, 24, 25, 26, 28, 29, 30, 39], "displai": [39, 41], "dissip": [2, 3, 4, 6, 8, 28, 29, 30], "distanc": [17, 19, 20, 21, 22, 39], "distinguish": 41, "distribut": [39, 40, 41], "do": [10, 35, 36, 37, 41], "document": [10, 16, 20, 37, 41], "doe": [6, 41], "domain": 41, "done": 36, "doubt": 41, "down": 36, "drela": [16, 28], "dstar": [20, 22], "dtype": [2, 3, 4, 9, 18, 19, 32], "du_e": [6, 8, 13, 14, 15, 17, 18, 19, 28, 29, 30], "du_e_smooth": 17, "du_edx": [6, 8, 28, 29, 30], "dui": 19, "dump": [16, 20, 21, 22], "durabl": 41, "dure": 4, "dwell": 41, "dx": [29, 30], "dynam": 41, "e": [10, 35], "each": [17, 18, 19, 20, 21, 22, 41], "earli": 11, "earlier": 41, "easier": 36, "econom": [35, 36], "edg": [2, 3, 4, 6, 8, 13, 14, 15, 17, 18, 19, 20, 21, 22, 28, 29, 30, 39], "edit": 35, "edland": [28, 29, 30, 39], "edu": 16, "educ": [35, 36], "effect": [39, 41], "effort": 41, "either": [2, 3, 4, 10, 11, 32, 39, 40, 41], "electron": [35, 41], "els": 39, "email": 36, "embodi": 41, "empathi": 35, "emploi": 6, "employ": 41, "empti": [20, 22], "en": 19, "enabl": 41, "encapsul": 8, "end": [6, 8, 9, 11, 28, 29, 30, 41], "energi": [2, 3, 4, 6, 8, 13, 14, 15, 17, 19, 20, 21, 28, 29, 30], "enforc": 41, "english": [17, 19], "enlish": 18, "ensur": 41, "enter": 41, "entir": [17, 41], "entiti": 41, "entrain": 39, "enumer": 9, "environ": [35, 36], "equal": [11, 19, 20, 22], "equat": [1, 2, 3, 4, 5, 8, 16, 27, 29, 30, 33, 39], "equilibrium": [17, 19], "equival": 41, "erron": 41, "error": 36, "essenti": 41, "eta": [2, 3, 4], "eta_": [2, 3, 4], "eta_d": [2, 3, 4], "eta_inf": [2, 3, 4], "eta_inf_default": 3, "eta_k": [2, 3, 4], "eta_m": [2, 3, 4], "ethnic": [35, 36], "even": [39, 40, 41], "event": [6, 8, 10, 28, 29, 30, 35, 41], "event_info": 10, "ever": 41, "everi": 41, "everyon": [35, 36, 41], "exact": [30, 41], "exactli": 11, "exampl": [11, 35, 36, 41], "except": 41, "exclud": 41, "exclus": 41, "excus": 41, "execut": 41, "exercis": 41, "exist": 10, "expect": [8, 35, 41], "experi": [35, 36], "experiment": [16, 17, 24, 25, 26], "explain": 41, "explan": 28, "explicit": 35, "explicitli": [8, 41], "express": [35, 36, 41], "expressli": 41, "extend": 41, "extens": 41, "extent": 41, "extern": 16, "extra": 10, "f": [2, 3, 4], "f_end": 9, "f_p": [2, 3, 4], "f_pp": [2, 3, 4], "face": 35, "facil": 41, "factor": [2, 3, 4, 6, 8, 13, 14, 15, 17, 19, 20, 21, 22, 24, 25, 26, 28, 29, 30, 39], "fail": 41, "failur": 41, "fair": [35, 41], "faith": 35, "falkner": [4, 13], "fals": [9, 11], "famili": 41, "faq": 35, "fashion": 41, "favor": 41, "featur": [8, 39, 41], "fee": 41, "feedback": 10, "few": [28, 29, 30, 39], "file": [16, 20, 21, 22, 41], "filenam": 21, "final": 41, "find": 41, "finit": [8, 28, 29, 30], "first": [6, 8, 28, 29, 30, 41], "fit": [27, 28, 29, 30, 39, 40, 41], "fix": [39, 41], "flag": [9, 18, 19], "float": [2, 3, 4, 6, 8, 9, 10, 13, 14, 15, 17, 19, 20, 21, 22, 28, 29, 30, 32], "flow": [1, 8, 13, 16, 17, 19, 21, 30, 39, 41], "focus": 35, "follow": [28, 29, 30, 35, 36, 41], "forbid": 41, "forc": 41, "form": 39, "format": 41, "formula": 19, "foster": [35, 36], "found": [2, 3, 4, 8, 41], "foundat": [39, 40, 41], "four": 17, "frac": [29, 30], "free": [35, 36, 39, 40, 41], "freestream": [6, 8, 28, 29, 30], "friction": [6, 17, 19, 20, 21, 23, 24, 25, 26], "from": [2, 3, 4, 6, 7, 8, 9, 10, 12, 13, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 28, 29, 30, 35, 36, 39], "from_byt": 11, "fsf": 41, "fulfil": 41, "full": 41, "function": [2, 3, 4, 6, 8, 23, 24, 25, 26, 28, 29, 30, 41], "fundament": 41, "further": [35, 41], "futur": 41, "fw_pp": [2, 3, 4], "fw_pp_default": 3, "g": 19, "gamma": 2, "gender": [35, 36], "gener": [2, 8, 39, 40], "get": [39, 41], "github": [36, 39], "give": 41, "given": [11, 17, 24, 25, 26, 39, 41], "gnu": [39, 40], "good": 35, "govern": 41, "gpl": 41, "gracefulli": 35, "grant": 41, "grati": 41, "greater": [4, 6, 8, 13, 14, 15, 19, 20, 21, 22, 28, 29, 30], "greatest": 41, "guarante": 41, "gui": 41, "h": [19, 20, 22, 30], "ha": [6, 7, 8, 9, 10, 41], "had": 41, "harass": [35, 36], "harm": 35, "have": [6, 8, 10, 24, 25, 26, 28, 29, 30, 35, 36, 39], "head": [5, 6, 39], "head1958": 39, "head_method": 39, "headmethod": 39, "healthi": 35, "help": [10, 37], "hereaft": 41, "high": 16, "hm": 39, "hold": 11, "holder": 41, "home": 39, "hope": [39, 40, 41], "host": [11, 41], "household": 41, "how": [39, 40], "howev": 41, "html": [35, 41], "http": [16, 35, 39, 41], "hypothet": 41, "i": [0, 2, 3, 4, 6, 8, 9, 10, 11, 14, 16, 17, 18, 19, 20, 21, 22, 24, 25, 26, 27, 28, 29, 30, 35, 36, 37, 39, 40, 41], "ibl": 36, "iblmethod": [6, 7, 9, 10, 12, 28, 29, 30], "ic": [6, 8, 28, 29, 30], "id": 36, "idea": 41, "ident": [35, 36], "identifi": 41, "idx": 17, "ifp": [16, 17], "ii": [16, 17], "imag": 11, "imageri": 35, "imaginari": 11, "implement": [0, 2, 5, 8, 10, 27, 28, 29, 30, 39, 41], "impli": [39, 40, 41], "import": [9, 39, 41], "impos": 41, "improv": [37, 39], "inabl": 41, "inaccur": 41, "inappropri": 35, "inc": 41, "incid": 35, "incident": 41, "includ": [16, 28, 35, 41], "inclus": [35, 41], "incompat": 41, "incompress": 2, "incorpor": 41, "indemnif": 41, "independ": 41, "index": [17, 21, 39], "indic": [6, 9, 10, 11, 18, 19, 41], "individu": [35, 41], "industri": 41, "inf": 9, "info": 39, "inform": [6, 8, 9, 10, 16, 17, 18, 28, 29, 30, 31, 35, 36, 39, 41], "infrastructur": 39, "infring": 41, "inher": 12, "inherit": 7, "initi": [2, 3, 4, 6, 8, 9, 12, 13, 14, 15, 17, 19, 20, 21, 22, 28, 29, 30, 39, 41], "initial_delta_m": [6, 28, 29, 30, 39], "initial_shape_d": [6, 39], "initialcondit": [6, 8, 12, 13, 15, 28, 29, 30], "inputparam": [6, 8, 9, 24, 25, 26, 28, 29, 30], "insid": 41, "instabl": [2, 3, 4], "instal": [36, 41], "instanc": [7, 35], "instead": 41, "insult": 35, "int": [11, 17, 18, 19, 21], "intact": 41, "intefac": 8, "integ": 11, "integr": [0, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 14, 16, 27, 28, 29, 30, 39], "integrationresult": [6, 7, 8, 28, 29, 30], "intend": [14, 39, 41], "intengr": 8, "intent": 41, "intenum": 11, "interact": 41, "interchang": 41, "interest": [2, 3, 4, 35, 36, 41], "interf": 41, "interfac": [12, 14, 16, 17, 21, 39, 41], "intern": [28, 29, 30], "interpol": [8, 18], "interst": 37, "intim": 41, "invalid": [2, 3, 4, 6, 8, 18, 19, 20, 22, 28, 29, 30, 41], "investig": 35, "inviscid": [2, 3, 4, 6, 8, 21, 28, 29, 30, 39], "irrevoc": 41, "issu": [35, 36], "item": 41, "iter": 11, "its": [6, 8, 16, 28, 29, 30, 35, 41], "itself": 41, "ix": [17, 18, 19], "journal": [28, 39], "june": 41, "k": 20, "ke_defect": 20, "ke_defect_low": 21, "ke_defect_upp": 21, "keep": 41, "kei": 41, "kernel": 41, "kind": 41, "kinemat": [2, 3, 4, 6, 8, 13, 14, 15, 17, 19, 28, 29, 30, 39], "kinet": [2, 3, 4, 6, 8, 13, 14, 15, 17, 19, 20, 21, 28, 29, 30], "know": [36, 41], "knowingli": 41, "knowledg": 41, "known": 8, "lambda": [29, 30], "laminar": [8, 11, 28, 29, 30, 39], "languag": [35, 41], "larger": 41, "last": 39, "later": [39, 40, 41], "lawsuit": 41, "layer": [0, 1, 2, 3, 4, 5, 6, 7, 8, 11, 12, 14, 16, 17, 19, 21, 23, 24, 25, 26, 27, 28, 29, 30, 39], "lead": 21, "leadership": 35, "least": [8, 41], "left": [29, 30], "length": [11, 21, 39], "lesser": 41, "let": 36, "level": [35, 36], "lgpl": 41, "liabl": 41, "librari": [31, 37, 39, 41], "license": 41, "licensor": 41, "like": [6, 8, 24, 26, 28, 29, 30, 36, 41], "likewis": 41, "line": 41, "linear": [27, 28, 29, 30], "link": 41, "list": [6, 8, 18, 19, 28, 29, 30, 41], "litig": 41, "littl": 11, "loation": [6, 8, 28, 29, 30], "local": [17, 19, 41], "locat": [2, 3, 4, 6, 8, 9, 17, 18, 19, 20, 21, 22, 28, 29, 30, 36], "long": [9, 41], "longer": 9, "loss": 41, "lot": 7, "lower": 21, "lower_count": 21, "lowest": 11, "ludwieg": [17, 19, 25], "m": [4, 17, 18, 19, 20, 21, 39], "machin": 41, "made": 41, "mai": [8, 35, 41], "mail": [35, 41], "main": [36, 39], "maintain": [35, 36, 41], "major": 41, "make": [35, 36, 41], "malachi": 39, "manner": 41, "manual": 15, "manufactur": 41, "march": 41, "mark": [16, 41], "marshal": 39, "mask": 41, "mass": [20, 21], "mass_defect": 20, "mass_defect_low": 21, "mass_defect_upp": 21, "materi": 41, "maximum": [2, 3, 4], "mean": 41, "measur": [17, 41], "mechan": 6, "media": 35, "medium": 41, "meet": 41, "member": 35, "menu": 41, "merchant": [39, 40, 41], "mere": 41, "merg": 41, "messag": [6, 8, 9, 10, 28, 29, 30, 36, 39], "met": [16, 41], "method": [0, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 27, 28, 29, 30, 39, 41], "michel": 39, "might": [39, 41], "minim": 36, "ministri": 39, "minor": 39, "misrepresent": 41, "mit": 16, "mixtur": 17, "mode": 41, "model": [6, 21, 28, 29, 30, 39, 41], "modern": 28, "modif": 41, "modifi": [39, 40], "modul": [0, 1, 5, 7, 8, 16, 23, 27, 31, 39], "mom_defect": 20, "mom_defect_low": 21, "mom_defect_upp": 21, "momentum": [2, 3, 4, 6, 8, 13, 14, 15, 17, 19, 20, 21, 22, 23, 24, 25, 26, 28, 29, 30, 39], "monoton": 8, "more": [8, 10, 16, 30, 37, 39, 40, 41], "moreov": 41, "most": [9, 11, 36, 41], "mostli": 39, "much": [7, 9, 36], "must": [2, 3, 4, 6, 8, 11, 13, 14, 15, 19, 20, 21, 22, 28, 29, 30, 41], "n_tran": 21, "name": [16, 21, 28, 29, 30, 41], "nation": [35, 36], "nativ": 11, "natur": 41, "ndarrai": [2, 3, 4, 6, 8, 9, 17, 18, 19, 21, 28, 29, 30, 32], "necessari": [5, 11, 27, 35, 41], "need": [2, 3, 4, 6, 8, 10, 12, 16, 28, 39, 41], "neeed": 31, "neg": 11, "neither": [8, 41], "network": 41, "new": [6, 20, 22, 39, 40], "next": 41, "non": [2, 3, 4, 17, 19], "noncommerci": 41, "nondimension": [20, 21, 22], "none": [2, 3, 4, 6, 8, 10, 14, 17, 18, 19, 20, 21, 22, 28, 29, 30], "nonlinear": 28, "nor": 41, "normal": [2, 3, 4, 20, 21, 22, 41], "note": [2, 3, 4, 10, 17, 18, 19, 20, 28], "noth": 41, "notic": 41, "notifi": 41, "notwithstand": 41, "np": 9, "nu": [6, 8, 13, 14, 15, 17, 19, 28, 29, 30], "nu_inf": 39, "nu_ref": [2, 3, 4, 17], "number": [6, 8, 11, 16, 17, 18, 19, 21, 23, 24, 25, 26, 28, 29, 30, 41], "numer": [11, 16], "numpi": [2, 3, 4, 6, 8, 17, 21, 25, 28, 29, 30, 32], "object": [8, 9, 11, 17, 18, 19, 20, 21, 22, 28, 29, 30, 41], "oblig": [35, 41], "obtain": [2, 3, 4, 13, 15, 18, 21, 39], "occasion": 41, "occur": [6, 41], "od": [2, 3, 4, 6, 8, 27, 28, 29, 30, 33], "offens": 35, "offer": 41, "offici": [35, 41], "offlin": 35, "olymp": [16, 17, 18, 19], "onc": [2, 3, 4], "one": [8, 27, 28, 29, 30, 41], "onli": [2, 3, 4, 41], "onlin": 35, "open": [35, 36], "oper": 41, "option": [2, 3, 4, 6, 8, 9, 17, 21, 28, 29, 30, 39, 40, 41], "order": [8, 11, 12, 35, 41], "ordinari": 8, "org": [35, 39, 41], "organ": 41, "orient": [35, 36], "origin": [11, 17, 19, 20, 22, 28, 29, 30, 39, 41], "other": [6, 8, 10, 35, 39], "otherwis": [2, 3, 4, 8, 35, 39, 41], "our": [36, 41], "out": [12, 39, 41], "output": [2, 3, 4, 8, 41], "outsid": [2, 3, 4, 41], "overflowerror": 11, "overrid": [2, 3, 4], "own": 41, "p": 20, "packag": [0, 9, 41], "page": [17, 18, 19, 39], "pair": 11, "panel": 16, "paper": [28, 29, 41], "paragraph": 41, "paramet": [2, 3, 4, 6, 8, 9, 10, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 24, 25, 26, 28, 29, 30, 39], "pars": 21, "part": [2, 3, 4, 11, 41], "parti": 41, "particip": [35, 36], "particular": [6, 8, 10, 17, 28, 29, 30, 39, 40, 41], "pass": [2, 3, 4, 8, 32, 41], "password": 41, "pattern": 41, "payment": 41, "pde": 3, "peer": 41, "pep": 36, "perform": [4, 41], "perman": [35, 41], "permiss": 35, "permit": 41, "perpetu": 41, "person": [35, 36, 41], "pertin": 41, "physic": [35, 41], "piec": 41, "place": 41, "plan": 36, "pleas": [36, 41], "pledg": 36, "plu": 41, "point": [2, 3, 4, 8, 11, 13, 18, 20, 21, 22, 39], "pointer": 41, "polici": 35, "polit": 35, "polynomi": 8, "portion": [21, 41], "posit": [2, 3, 4, 11, 35], "possess": 41, "possibl": [7, 9, 36, 41], "post": 35, "power": 41, "pp": [28, 39], "practic": 41, "preambl": [39, 40], "precis": [17, 18, 19, 41], "predecessor": 41, "predict": [6, 16], "prefer": 41, "present": 41, "preserv": 41, "prevent": 41, "previou": 41, "price": 41, "primarili": 41, "print": 39, "prior": 41, "privat": [35, 41], "problem": [36, 41], "procedur": 41, "proceed": [16, 17, 18, 19], "process": [2, 3, 4, 6, 8, 9, 11, 21, 28, 29, 30], "procur": 41, "produc": [11, 36, 41], "product": 41, "profession": 35, "profil": [4, 6, 13, 14, 15, 18, 29, 30, 39], "program": [39, 40], "programm": 41, "prohibit": 41, "project": [35, 36, 39], "promin": 41, "propag": 41, "properti": [2, 3, 4, 6, 8, 9, 13, 14, 15, 17, 18, 19, 20, 21, 22, 28, 29, 30, 39, 41], "proprietari": 41, "protocol": [11, 41], "prove": 41, "provid": [1, 6, 7, 8, 9, 10, 12, 14, 15, 16, 17, 23, 27, 28, 29, 30, 35, 36, 39, 41], "provis": 41, "provision": 41, "proxi": 41, "public": [16, 35, 39, 40], "publicli": 41, "publish": [35, 39, 40, 41], "purpos": [10, 39, 40, 41], "pursuant": 41, "pycodestyl": 36, "pylint": 36, "pypi": 36, "python": [0, 39], "qualifi": 41, "qualiti": [16, 37, 41], "queri": [8, 39], "question": 35, "r": 39, "race": [35, 36], "radian": 21, "rais": [2, 3, 4, 6, 8, 11, 18, 19, 20, 22, 28, 29, 30], "rang": [2, 3, 4], "rate": [8, 13, 14, 15, 17, 18, 19], "ratio": 11, "ration": 11, "rdel": 19, "re_delta_d": [17, 19], "re_delta_m": [17, 19, 24, 25, 26], "reach": [6, 8, 28, 29, 30], "reached_end": 11, "read": [16, 21, 41], "readabl": 41, "reader": 21, "readi": 41, "readili": 41, "real": 11, "reason": [9, 10, 11, 35, 41], "receipt": 41, "receiv": [39, 41], "recogn": 41, "redistribut": [39, 40, 41], "refactor": 39, "refer": [2, 3, 4, 28, 39, 41], "refrain": 41, "regard": [35, 41], "regardless": [35, 36, 41], "regener": 41, "reinstat": 41, "reject": 35, "rel": 39, "relat": [6, 8, 10, 24, 25, 26, 28, 29, 30], "relationship": [27, 41], "releas": [39, 41], "relev": 41, "reli": [28, 41], "relicens": 41, "religion": [35, 36], "remain": 41, "remov": [27, 35, 41], "render": 41, "repair": 41, "repercuss": 35, "report": [2, 3, 4, 16, 17, 19, 20, 22, 35], "repositori": 36, "repres": [2, 3, 4, 9, 11, 28, 29, 30, 35, 41], "represent": [6, 8, 11, 17, 28, 29, 30, 35], "reproduc": 36, "request": 11, "requir": [2, 3, 4], "research": 39, "reset": [18, 19, 20, 22], "reset_beta": 4, "reset_m": 4, "resolv": 41, "resourc": [16, 39], "respect": [35, 41], "respons": 41, "restrict": 41, "result": [8, 9, 16, 27, 35, 39, 41], "retain": 41, "return": [2, 3, 4, 6, 7, 8, 9, 10, 11, 13, 14, 15, 17, 18, 19, 20, 21, 22, 24, 25, 26, 28, 29, 30, 41], "review": [35, 41], "reynold": [17, 19, 21, 23, 24, 25, 26], "rho": [6, 8, 28, 29, 30], "rho_inf": 39, "rho_ref": [2, 3, 4], "right": [29, 30, 35], "risk": 41, "rom": 41, "root": 36, "row": [19, 20, 22], "royalti": 41, "rtheta": 19, "rtn": 39, "rule": 41, "run": [6, 8, 28, 29, 30, 41], "s_lower": 21, "s_upper": 21, "s_wake": 21, "safest": 41, "sai": 41, "sake": 41, "sale": 41, "same": [2, 3, 4, 8, 24, 25, 26, 39, 41], "sampl": [17, 18, 19], "sample_count": 19, "satisfi": 41, "scalar": [6, 8, 28, 29, 30], "scalartyp": [2, 3, 4, 9, 18, 19, 32], "school": 41, "scipi": [8, 9], "scope": 41, "script": 41, "search": [4, 39], "second": [6, 8, 28, 29, 30], "secondarili": 41, "see": [6, 8, 10, 16, 17, 18, 19, 20, 28, 29, 30, 35, 36, 39, 40, 41], "self": 11, "sell": 41, "semiconductor": 41, "separ": [6, 11, 21, 35, 41], "serv": 41, "server": 41, "servic": 41, "set": [2, 3, 4, 6, 8, 9, 12, 15, 16, 28, 29, 30, 35, 39], "set_data_fit": [28, 29, 30], "set_initial_condit": [6, 8, 28, 29, 30], "set_shape_d_crit": 6, "set_solution_paramet": [2, 3, 4], "set_veloc": [6, 8, 28, 29, 30], "sex": [35, 36], "sexual": [35, 36], "shall": 41, "shape": [2, 3, 4, 6, 8, 13, 14, 15, 17, 19, 20, 21, 22, 24, 25, 26, 28, 29, 30, 39], "shape_d": [2, 3, 4, 6, 8, 13, 14, 15, 17, 19, 20, 22, 24, 25, 26, 28, 29, 30], "shape_d0": 39, "shape_d_crit": 6, "shape_d_low": 21, "shape_d_upp": 21, "shape_d_wak": 21, "shape_eq": [17, 19], "shape_k": [2, 3, 4, 6, 8, 13, 14, 15, 17, 19, 20, 28, 29, 30], "shape_k_low": 21, "shape_k_upp": 21, "share": 41, "shear": [2, 3, 4, 6, 8, 17, 19, 28, 29, 30, 39], "short": 41, "should": [2, 3, 4, 7, 10, 36, 39, 41], "show": [35, 41], "si": [17, 18, 19], "si_unit": [18, 19], "sign": [10, 11, 41], "signific": [11, 41], "similar": [2, 3, 4, 41], "similarli": 39, "simplifi": 39, "simultan": 41, "sinc": [6, 28], "singl": [39, 41], "situat": 10, "size": [18, 35, 36], "skan": [4, 13], "skin": [6, 17, 19, 20, 21, 23, 24, 25, 26], "slightli": 27, "slope": [28, 29, 30], "small": 16, "smooth": [17, 18], "so": 41, "social": 35, "socio": [35, 36], "softwar": [39, 40, 41], "sold": 41, "sole": 41, "solut": [1, 2, 3, 4, 6, 8, 12, 13, 14, 16, 21, 28, 29, 30, 33, 39], "solv": [2, 3, 4, 6, 8, 10, 27, 28, 29, 30, 39], "solve_ivp": [8, 9, 10], "solver": [2, 3, 4, 6, 7, 8, 9, 10, 16, 28, 29, 30], "some": [39, 41], "space": 35, "spare": 41, "speak": 41, "special": 41, "specif": [2, 3, 4, 6, 8, 10, 19, 28, 29, 30, 35, 41], "specifi": [2, 3, 4, 6, 8, 17, 21, 28, 29, 30, 41], "spirit": 41, "spline": [8, 28, 29, 30], "stagnat": [8, 13, 15, 20, 21, 22, 39], "stand": 41, "standard": 41, "stanford": [16, 17, 18, 19], "start": [6, 8, 12, 28, 29, 30, 41], "stat_data": 19, "stat_summ": 19, "state": [9, 41], "statement": 41, "station": [17, 18, 19, 20, 22], "station_count": 17, "stationdata": 17, "statu": [9, 35, 36, 41], "step": 41, "storag": 41, "store": [8, 21], "str": [9, 10, 17, 18, 19, 20, 21, 22, 28, 29, 30], "stream": [2, 3, 4], "streamwis": [2, 3, 4, 6, 8, 17, 18, 19, 28, 29, 30], "stress": [2, 3, 4, 6, 8, 17, 28, 29, 30, 39], "string": [17, 19, 20, 22, 28, 29, 30], "subdivid": 41, "subject": 41, "sublicens": 41, "subprogram": 41, "subroutin": 41, "subsect": 41, "substanti": 41, "success": [9, 39], "successfulli": 9, "sue": 41, "suffic": 41, "suffici": 3, "summ_data": 19, "summari": 19, "supplement": 41, "support": [11, 36, 39, 41], "sure": 41, "surfac": [17, 19, 20, 21, 22, 39], "surviv": 41, "sustain": 41, "sy": 11, "symbol": [19, 20, 22], "system": [6, 11, 33, 41], "systemat": 41, "tabl": 19, "tabul": 28, "tabular": [28, 29, 30], "take": [28, 29, 30, 35, 39, 41], "taken": 8, "tangibl": 41, "tau_w": [2, 3, 4, 6, 8, 28, 29, 30, 39], "tau_wal": 39, "tbd": 36, "team": 35, "technolog": 41, "tell": 41, "temporari": 35, "temporarili": 35, "term": [11, 17, 18, 19, 20, 39, 40], "term_ev": [6, 8, 28, 29, 30], "termev": [6, 8, 28, 29, 30], "termin": [6, 8, 9, 10, 28, 29, 30], "termreason": [9, 10], "test": [27, 39], "th": 19, "than": [4, 6, 8, 13, 14, 15, 19, 20, 21, 22, 28, 29, 30, 41], "thei": [12, 24, 25, 26, 35, 36, 41], "them": 41, "therefor": 41, "thesi": 39, "theta": [19, 20, 22], "thi": [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 35, 36, 37, 39, 40], "thick": [2, 3, 4, 6, 8, 13, 14, 15, 17, 19, 20, 21, 22, 23, 24, 25, 26, 28, 29, 30, 39], "thin": 39, "thing": 41, "third": 41, "those": [37, 41], "though": 41, "threaten": [35, 41], "three": [28, 29, 30, 41], "threshold": 6, "through": 41, "throughout": 17, "thu": 41, "thwait": [27, 28, 29, 30, 39], "thwaites1949": 39, "thwaites_method": 39, "thwaitesmethod": [29, 30], "thwaitesmethodlinear": [27, 28], "thwaitesmethodnonlinear": [27, 28, 39], "tightli": 16, "tillman": [17, 19, 25], "time": 41, "tm": 39, "to_byt": 11, "too": 41, "tool": [36, 41], "toward": 35, "track": 36, "tracker": 36, "trade": 41, "trademark": 41, "tradit": 27, "trail": 21, "transact": 41, "transfer": 41, "transit": [11, 21, 39], "transmiss": 41, "transpir": [2, 3, 4, 6, 8, 28, 29, 30, 39], "treat": 41, "treati": 41, "trigger": [6, 8, 28, 29, 30], "troll": 35, "true": 10, "tupl": [6, 8, 10, 28, 29, 30], "turbul": [6, 16, 17, 19, 23, 24, 25, 26, 39], "two": [5, 10, 11, 16, 27, 41], "type": [2, 3, 4, 6, 8, 10, 13, 14, 15, 17, 18, 19, 20, 21, 22, 24, 25, 26, 28, 29, 30, 41], "typeerror": [6, 8, 28, 29, 30], "typic": 41, "u": [2, 3, 4, 17, 19, 36], "u_": [2, 3, 4, 6, 8, 17, 18, 19, 28, 29, 30, 39], "u_defect": [17, 19], "u_e_low": 21, "u_e_rel": [20, 22], "u_e_smooth": 17, "u_e_upp": 21, "u_e_wak": 21, "u_plu": [17, 19], "u_ref": [2, 3, 4, 21], "u_star": [17, 19], "udef": 19, "ue": [20, 22], "ui": 19, "unaccept": [35, 41], "under": [39, 40, 41], "uniformli": 8, "unit": [17, 18, 19, 36], "unknown": [9, 10, 11], "unless": 41, "unlimit": 41, "unmodifi": 41, "unnecessari": 41, "unpack": 41, "until": 41, "unus": [28, 29, 30], "unwelcom": 35, "up": [6, 8, 28, 29, 30, 39], "updat": [39, 41], "upon": 28, "upper": 21, "upper_count": 21, "us": [2, 3, 4, 6, 8, 10, 11, 12, 16, 17, 18, 19, 20, 21, 22, 24, 25, 26, 28, 29, 30, 31, 32, 33, 35, 36, 37, 39, 40], "user": [3, 6, 8, 10, 28, 29, 30], "v": [2, 3, 4, 19], "v_e": [2, 3, 4, 6, 8, 28, 29, 30], "valid": [2, 3, 4, 10, 41], "valu": [2, 3, 4, 6, 8, 9, 10, 11, 20, 22, 28], "valueerror": [2, 3, 4, 6, 8, 18, 19, 20, 22, 28, 29, 30], "variat": 8, "varieti": [1, 4, 16, 17, 39], "variou": 7, "veloc": [2, 3, 4, 6, 8, 13, 14, 15, 17, 18, 19, 20, 21, 22, 28, 29, 30, 39], "veri": 8, "version": [28, 35, 40], "via": [28, 29, 30, 35, 36], "view": 41, "viewpoint": 35, "vinf": [20, 22], "violat": 41, "viscos": [2, 3, 4, 6, 8, 13, 14, 15, 17, 19, 28, 29, 30, 39], "viscou": 39, "visibl": 41, "void": 41, "vol": [28, 39], "volum": [16, 17, 41], "w": 41, "wa": [6, 8, 9, 28, 29, 30, 41], "wai": [6, 8, 28, 29, 30, 41], "waiv": 41, "waiver": 41, "wake": [21, 22], "wake_count": 21, "wall": [2, 3, 4, 6, 8, 17, 19, 28, 29, 30, 39], "want": [17, 41], "warn": 36, "warranti": [39, 40], "we": [35, 36, 41], "web": 16, "wedg": 4, "welcom": [35, 36, 41], "well": 41, "were": [17, 41], "what": [10, 35, 36, 41], "whatev": 41, "when": [6, 7, 8, 10, 28, 29, 30, 32, 35, 36, 41], "where": 41, "whether": [9, 11, 18, 19, 41], "which": [35, 39, 41], "while": [3, 17], "white": [28, 29, 30], "who": [35, 37, 41], "whole": 41, "whom": 41, "whose": [11, 41], "why": [10, 41], "wide": [4, 16, 41], "wiki": 35, "window": 41, "wipo": 41, "wish": 41, "within": [9, 10, 35, 41], "without": [28, 30, 35, 39, 40, 41], "work": [29, 30, 36, 39, 41], "worldwid": 41, "would": [10, 36, 41], "write": 41, "written": [16, 41], "www": [35, 39, 41], "x": [2, 3, 4, 6, 8, 17, 18, 19, 20, 22, 28, 29, 30], "x0": [6, 8, 28, 29, 30, 39], "x_end": [6, 8, 9, 28, 29, 30, 39], "x_lower": 21, "x_smooth": 17, "x_trans_low": 21, "x_trans_upp": 21, "x_upper": 21, "x_wake": 21, "xfoil": [16, 20, 21, 22], "xpoint": 8, "y": [2, 3, 4, 17, 19, 20, 22], "y_c": [17, 19], "y_lower": 21, "y_plu": [17, 19], "y_upper": 21, "y_wake": 21, "year": 41, "you": [36, 39, 40, 41], "your": [36, 39, 40], "yourself": 41, "zero": [6, 8, 13, 14, 15, 19, 20, 21, 22, 28, 29, 30]}, "titles": ["ibl", "ibl.analytic", "ibl.analytic.Analytic2dSimilarityIncompressible", "ibl.analytic.Blasius", "ibl.analytic.FalknerSkan", "ibl.head_method", "ibl.head_method.HeadMethod", "ibl.ibl_method", "ibl.ibl_method.IBLMethod", "ibl.ibl_method.IntegrationResult", "ibl.ibl_method.TermEvent", "ibl.ibl_method.TermReason", "ibl.initial_condition", "ibl.initial_condition.FalknerSkanStagCondition", "ibl.initial_condition.InitialCondition", "ibl.initial_condition.ManualCondition", "ibl.reference", "ibl.reference.StanfordOlympics1968", "ibl.reference.StanfordOlympics1968SmoothVel", "ibl.reference.StanfordOlympics1968StationData", "ibl.reference.XFoilAirfoilData", "ibl.reference.XFoilReader", "ibl.reference.XFoilWakeData", "ibl.skin_friction", "ibl.skin_friction.felsch", "ibl.skin_friction.ludwieg_tillman", "ibl.skin_friction.white", "ibl.thwaites_method", "ibl.thwaites_method.ThwaitesMethod", "ibl.thwaites_method.ThwaitesMethodLinear", "ibl.thwaites_method.ThwaitesMethodNonlinear", "ibl.typing", "ibl.typing.InputParam", "ibl.typing.SolutionFunc", "ibl", "Contributor Covenant Code of Conduct", "Contributing", "Developer Information", "Examples", "Welcome to IBL\u2019s documentation!", "License Information", "GNU GENERAL PUBLIC LICENSE"], "titleterms": {"": 39, "0": 41, "1": 41, "10": 41, "11": 41, "12": 41, "13": 41, "14": 41, "15": 41, "16": 41, "17": 41, "2": 41, "3": 41, "4": 41, "5": 41, "6": 41, "7": 41, "8": 41, "9": 41, "AND": 41, "No": 41, "Not": 41, "These": 41, "accept": 41, "addit": 41, "affero": 41, "analyt": [1, 2, 3, 4], "analytic2dsimilarityincompress": 2, "anti": 41, "appli": 41, "attribut": 35, "automat": 41, "basic": 41, "blasiu": 3, "bug": 36, "c": 36, "chang": 36, "circumvent": 41, "cmake": 36, "code": [35, 36, 41], "condit": 41, "conduct": [35, 36], "content": 39, "contribut": 36, "contributor": [35, 39], "convei": 41, "copi": 41, "coven": 35, "definit": 41, "develop": 37, "disclaim": 41, "document": 39, "downstream": 41, "enforc": 35, "exampl": [38, 39], "falknerskan": 4, "falknerskanstagcondit": 13, "felsch": 24, "file": 36, "form": 41, "freedom": 41, "from": 41, "gener": 41, "gnu": 41, "guid": 36, "have": 41, "head_method": [5, 6], "headmethod": 6, "histori": 39, "how": 41, "ibl": [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 39], "ibl_method": [7, 8, 9, 10, 11], "iblmethod": 8, "indic": 39, "inform": [37, 40], "initial_condit": [12, 13, 14, 15], "initialcondit": 14, "inputparam": 32, "integrationresult": 9, "interpret": 41, "latex": 36, "law": 41, "legal": 41, "liabil": 41, "licens": [39, 40, 41], "limit": 41, "ludwieg_tillman": 25, "manualcondit": 15, "matlab": 36, "modifi": 41, "new": 41, "non": 41, "other": [36, 41], "our": 35, "patent": 41, "permiss": 41, "pledg": 35, "preambl": 41, "program": 41, "protect": 41, "public": 41, "python": 36, "recipi": 41, "refer": [16, 17, 18, 19, 20, 21, 22], "report": 36, "requir": 41, "respons": 35, "revis": 41, "right": 41, "scope": 35, "script": 36, "section": 41, "skin_frict": [23, 24, 25, 26], "solutionfunc": 33, "sourc": 41, "standard": 35, "stanfordolympics1968": 17, "stanfordolympics1968smoothvel": 18, "stanfordolympics1968stationdata": 19, "style": 36, "submit": 36, "summari": 39, "surrend": 41, "tabl": 39, "term": 41, "termev": 10, "termin": 41, "termreason": 11, "test": 36, "thi": 41, "thwaites_method": [27, 28, 29, 30], "thwaitesmethod": 28, "thwaitesmethodlinear": 29, "thwaitesmethodnonlinear": 30, "type": [31, 32, 33], "us": 41, "usag": 39, "user": 41, "verbatim": 41, "version": [39, 41], "warranti": 41, "welcom": 39, "white": 26, "xfoilairfoildata": 20, "xfoilread": 21, "xfoilwakedata": 22, "your": 41}})
