@@ -1,3 +1,6 @@
+# pylint: skip-file
+# pyright: reportUntypedNamedTuple=false
+
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
@@ -6,12 +9,15 @@ Created on Fri Aug  5 22:56:30 2022
 @author: ddmarshall
 """
 
+from typing import Tuple
+
 import numpy as np
 import csv
 from collections import namedtuple
 
 
-def read_xfoil_dump_file(xfoil_file, U_inf=1):
+def read_xfoil_dump_file(xfoil_filename: str,
+                         U_inf: float = 1) -> Tuple[Tuple, Tuple, Tuple]:
     """
     Return the results from XFoil dump file.
     
@@ -35,10 +41,13 @@ def read_xfoil_dump_file(xfoil_file, U_inf=1):
         H_upper - upper surface shape factor
     """
     
-    PanelValues = namedtuple('PanelValues', ['s', 'x', 'y', 'Ue', 'delta_d', 'delta_m', 'cf', 'H', 'Hstar', 'P', 'm', 'K'])
-    WakeValues = namedtuple('PanelValues', ['s', 'x', 'y', 'Ue', 'delta_d', 'delta_m', 'H'])
+    PanelValues = namedtuple('PanelValues', ['s', 'x', 'y', 'Ue', 'delta_d',
+                                             'delta_m', 'cf', 'H', 'Hstar',
+                                             'P', 'm', 'K'])
+    WakeValues = namedtuple('WakeValues', ['s', 'x', 'y', 'Ue', 'delta_d',
+                                           'delta_m', 'H'])
     
-    xfoil_file = open(xfoil_file, 'r')
+    xfoil_file = open(xfoil_filename, 'r')
     xfoil_reader = csv.reader(xfoil_file)
     panel_terms = []
     wake_terms = []
@@ -54,19 +63,21 @@ def read_xfoil_dump_file(xfoil_file, U_inf=1):
                 raise Exception('Invalid number of columns for XFoil import')
                     
     ## Find stagnation point and the total arclength of airfoil
-    panel_terms = np.asarray(panel_terms)
-    last_upper_idx = np.where(np.sign(panel_terms[:-1, 3]) != np.sign(panel_terms[1:, 3]))[0][0]
-    s_af = panel_terms[-1, 0]
+    panel_array = np.asarray(panel_terms)
+    last_upper_idx = np.where(np.sign(panel_array[:-1, 3])
+                              != np.sign(panel_array[1:, 3]))[0][0]
+    s_af = panel_array[-1, 0]
     
     # interpolate stagnation point
-    dU = panel_terms[last_upper_idx+1, 3]-panel_terms[last_upper_idx, 3]
-    frac = -panel_terms[last_upper_idx, 3]/dU
-    stag_col = frac*panel_terms[last_upper_idx+1,:]+(1-frac)*panel_terms[last_upper_idx,:]
+    dU = panel_array[last_upper_idx+1, 3]-panel_array[last_upper_idx, 3]
+    frac = -panel_array[last_upper_idx, 3]/dU
+    stag_col = (frac*panel_array[last_upper_idx+1,:]
+                + (1 - frac)*panel_array[last_upper_idx,:])
 
     ## Split into upper and lower surface terms
     # upper panel needs to be reversed since XFoil writes out starting with 
     #   the upper surface trailing edge to lower surface trailing edge.
-    panel_upper_terms = panel_terms[last_upper_idx::-1, :]
+    panel_upper_terms = panel_array[last_upper_idx::-1, :]
     # insert the stagnation point
     panel_upper_terms = np.insert(panel_upper_terms, 0, stag_col, 0)
     # adjust distance is based on arclength distance from upper trailing edge.
@@ -75,7 +86,7 @@ def read_xfoil_dump_file(xfoil_file, U_inf=1):
     upper_values = PanelValues(*panel_upper_terms.T)
     
     # lower panel
-    panel_lower_terms = panel_terms[last_upper_idx+1:, :]
+    panel_lower_terms = panel_array[last_upper_idx+1:, :]
     # insert the stagnation point
     panel_lower_terms = np.insert(panel_lower_terms, 0, stag_col, 0)
     # adjust distance (see upper panel)
@@ -89,11 +100,11 @@ def read_xfoil_dump_file(xfoil_file, U_inf=1):
     
     # wake
     # adjust distance (base it on trailing edge)
-    wake_terms = np.asarray(wake_terms)
-    wake_terms[:,0] = wake_terms[:,0] - s_af
+    wake_array = np.asarray(wake_terms)
+    wake_array[:,0] = wake_array[:,0] - s_af
     # remove zeros from skin friction coefficient
-    wake_terms = np.delete(wake_terms, 6, 1)
-    wake_values = WakeValues(*wake_terms.T)
+    wake_array = np.delete(wake_array, 6, 1)
+    wake_values = WakeValues(*wake_array.T)
     
     return upper_values, lower_values, wake_values
 
