@@ -1,41 +1,40 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-Created on Tue Aug  9 17:26:49 2022
-
-@author: ddmarshall
-"""
-
+"""Module to test the base IBL functionality."""
 
 import unittest
-from typing import Tuple
+from typing import Tuple, Optional, Any
+from typing_extensions import override
+
 import numpy as np
-import numpy.testing as npt
-import numpy.typing as np_type
+import numpy.typing as npt
+import numpy.testing as np_type
+
 from scipy.interpolate import PchipInterpolator
 
-from pyBL.ibl_method import IBLMethod
-from pyBL.ibl_method import IBLTermEvent
-from pyBL.ibl_method import IBLResult
+from ibl.ibl_method import IBLMethod
+from ibl.ibl_method import TermEvent
+from ibl.ibl_method import TermReason
+from ibl.ibl_method import IntegrationResult
+from ibl.typing import InputParam
 
 
-class TestIBLResult(unittest.TestCase):
-    """Class to test IBLResult."""
+class TestIntegrationResult(unittest.TestCase):
+    """Class to test IntegrationResult."""
 
-    def test_string(self):
+    def test_string(self) -> None:
         """Test string representation."""
-        iblr = IBLResult(x_end=2.1, F_end=[1.4, -2.3], status=0,
-                         message="Success", success=True)
-        str_ref = ("IBLResult:\n"
+        iblr = IntegrationResult(x_end=2.1, f_end=np.array([1.4, -2.3]),
+                                 status=TermReason.REACHED_END,
+                                 message="Success", success=True)
+        str_ref = ("IntegrationResult:\n"
                    "    x_end: 2.1\n"
-                   "    F_end: [1.4, -2.3]\n"
-                   "    status: 0\n"
+                   "    f_end: [ 1.4 -2.3]\n"
+                   "    status: Completed\n"
                    "    message: Success\n"
                    "    success: True")
         self.assertEqual(str_ref, str(iblr))
 
 
-class _IBLMethodTestTermEvent(IBLTermEvent):
+class _TestTermEvent(TermEvent):
     """
     Sample class to test the termination capabilities of the base class.
 
@@ -47,11 +46,12 @@ class _IBLMethodTestTermEvent(IBLTermEvent):
         _x_kill: x-location that the integrator should stop.
     """
 
-    def __init__(self, x_kill):
+    def __init__(self, x_kill: float) -> None:
         self._x_kill = x_kill
         super().__init__()
 
-    def _call_impl(self, x, F):
+    @override
+    def _call_impl(self, x: float, f: InputParam) -> float:
         """
         Information used to determine if IBL test integrator should terminate.
 
@@ -60,8 +60,10 @@ class _IBLMethodTestTermEvent(IBLTermEvent):
 
         Parameters
         ----------
-            x: Current x-location of the integration
-            F: Current state value(s)
+            x : float
+                Current x-location of the integration
+            f : InputParam
+                Current state value(s)
 
         Returns
         -------
@@ -69,316 +71,605 @@ class _IBLMethodTestTermEvent(IBLTermEvent):
             the integration has passed the termination condition, and zero at
             the state when the integrator should terminate.
         """
+        _ = f
         return x - self._x_kill
 
-    def event_info(self):
-        return -1, ""
+    @override
+    def event_info(self) -> Tuple[TermReason, str]:
+        """
+        Return information about the purpose of this event.
+
+        This is used to provide feedback as to what caused the integration to
+        terminate and any other helpful information.
+
+        Returns
+        -------
+        TermReason
+            Value indicating reason this event would terminate integration
+        str
+            Extra information associated with reason for termination.
+
+        Notes
+        -----
+        The value `TermReason.UNKNOWN` should be used for reasons that do
+        not have an existing reason. For these termination situations the
+        message can provide more details.
+        """
+        return TermReason.SEPARATED, ""
 
 
 class IBLMethodTest(IBLMethod):
     """Generic class to test the concrete methods in IBLMethod"""
 
-    def __init__(self, U_e=None, dU_edx=None, d2U_edx2=None, x_kill=None):
+    def __init__(self, u_e: Any = None, du_e: Any = None, d2u_e:Any = None,
+                 x_kill: Optional[float] = None) -> None:
         # setup base class
-        super().__init__(nu=1, U_e=U_e, dU_edx=dU_edx, d2U_edx2=d2U_edx2)
+        super().__init__(nu=1, u_e=u_e, du_e=du_e, d2u_e=d2u_e)
         self.y0 = np.array([0])
 
         # set up this class
         if x_kill is not None:
-            self._set_kill_event(_IBLMethodTestTermEvent(x_kill))
+            self._set_kill_event(_TestTermEvent(x_kill))
 
-    def _ode_setup(self) -> Tuple[np_type.NDArray, float, float]:
+    @override
+    def _ode_setup(self) -> Tuple[npt.NDArray, Optional[float],
+                                  Optional[float]]:
         """
         Set the solver specific parameters.
 
         Returns
         -------
-        3-Tuple
-            IBL initialization array
-            Relative tolerance for ODE solver
-            Absolute tolerance for ODE solver
+        numpy.ndarray
+            IBL initialization array.
+        Optional[float]
+            Relative tolerance for ODE solver.
+        Optional[float]
+            Absolute tolerance for ODE solver.
         """
         return self.y0, None, None
 
-    def _ode_impl(self, x, F):
+    @override
+    def _ode_impl(self, x: npt.NDArray,
+                  f: npt.NDArray) -> npt.NDArray:
         """
         This is the derivatives of the ODEs that are to be solved
 
         Parameters
         ----------
-            x: x-location of current step
-            F: current step state value(s)
+            x : numpy.ndarray
+                x-location of current step
+            f : numpy.ndarray
+                current step state value(s)
+
+        Returns
+        -------
+        numpy.ndarray
+            The right-hand side of the ODE at the given state.
         """
-        _ = F
+        _ = f
         return x
 
-    def V_e(self, x):
+    @override
+    def v_e(self, x: InputParam) -> npt.NDArray:
+        """
+        Calculate the transpiration velocity.
+
+        Parameters
+        ----------
+        x: InputParam
+            Streamwise loations to calculate this property.
+
+        Returns
+        -------
+        numpy.ndarray
+            Desired transpiration velocity at the specified locations.
+        """
         return np.zeros_like(x)
 
-    def delta_d(self, x):
+    @override
+    def delta_d(self, x: InputParam) -> npt.NDArray:
+        """
+        Calculate the displacement thickness.
+
+        Parameters
+        ----------
+        x: InputParam
+            Streamwise loations to calculate this property.
+
+        Returns
+        -------
+        numpy.ndarray
+            Desired displacement thickness at the specified locations.
+        """
         return np.zeros_like(x)
 
-    def delta_m(self, x):
+    @override
+    def delta_m(self, x: InputParam) -> npt.NDArray:
+        """
+        Calculate the momentum thickness.
+
+        Parameters
+        ----------
+        x: InputParam
+            Streamwise loations to calculate this property.
+
+        Returns
+        -------
+        numpy.ndarray
+            Desired momentum thickness at the specified locations.
+        """
         return np.zeros_like(x)
 
-    def delta_k(self, x):
+    @override
+    def delta_k(self, x: InputParam) -> npt.NDArray:
+        """
+        Calculate the kinetic energy thickness.
+
+        Parameters
+        ----------
+        x: InputParam
+            Streamwise loations to calculate this property.
+
+        Returns
+        -------
+        numpy.ndarray
+            Desired kinetic energy thickness at the specified locations.
+        """
         return np.zeros_like(x)
 
-    def H_d(self, x):
+    @override
+    def shape_d(self, x: InputParam) -> npt.NDArray:
+        """
+        Calculate the displacement shape factor.
+
+        Parameters
+        ----------
+        x: InputParam
+            Streamwise loations to calculate this property.
+
+        Returns
+        -------
+        numpy.ndarray
+            Desired displacement shape factor at the specified locations.
+        """
         return np.zeros_like(x)
 
-    def H_k(self, x):
+    @override
+    def shape_k(self, x: InputParam) -> npt.NDArray:
+        """
+        Calculate the kinetic energy shape factor.
+
+        Parameters
+        ----------
+        x: InputParam
+            Streamwise loations to calculate this property.
+
+        Returns
+        -------
+        numpy.ndarray
+            Desired kinetic energy shape factor at the specified locations.
+        """
         return np.zeros_like(x)
 
-    def tau_w(self, x, rho):
+    @override
+    def tau_w(self, x: InputParam, rho: float) -> npt.NDArray:
+        """
+        Calculate the wall shear stress.
+
+        Parameters
+        ----------
+        x: InputParam
+            Streamwise loations to calculate this property.
+        rho: float
+            Freestream density.
+
+        Returns
+        -------
+        numpy.ndarray
+            Desired wall shear stress at the specified locations.
+        """
+        _ = rho
         return np.zeros_like(x)
 
-    def D(self, x, rho):
+    @override
+    def dissipation(self, x: InputParam, rho: float) -> npt.NDArray:
+        """
+        Calculate the dissipation integral.
+
+        Parameters
+        ----------
+        x: InputParam
+            Streamwise loations to calculate this property.
+        rho: float
+            Freestream density.
+
+        Returns
+        -------
+        numpy.ndarray
+            Desired dissipation integral at the specified locations.
+        """
+        _ = rho
         return np.zeros_like(x)
 
 
-class IBLMethodTestTransition(IBLTermEvent):
+class IBLMethodTestTransition(TermEvent):
     """Generic class to pass termination events during the solve method."""
 
-    def __init__(self, F_kill):
-        self._F_kill = F_kill
+    def __init__(self, f_kill: float) -> None:
+        self._f_kill = f_kill
         super().__init__()
 
-    def _call_impl(self, x, F):
-        return F[0] - self._F_kill
+    @override
+    def _call_impl(self, x: float, f: npt.NDArray) -> float:
+        """
+        Information used to determine if IBL test integrator should terminate.
 
-    def event_info(self):
-        return 1, ""
+        This will terminate once f passed specified value (f_kill) and will be
+        negative before then (positive afterwards).
+
+        Parameters
+        ----------
+            x : float
+                Current x-location of the integration
+            f : numpy.ndarray
+                Current state value(s)
+
+        Returns
+        -------
+        float
+            Negative value when the integration should continue, positive when
+            the integration has passed the termination condition, and zero at
+            the state when the integrator should terminate.
+        """
+        _ = x
+        return f[0] - self._f_kill
+
+    @override
+    def event_info(self) -> Tuple[TermReason, str]:
+        """
+        Return information about the purpose of this event.
+
+        This is used to provide feedback as to what caused the integration to
+        terminate and any other helpful information.
+
+        Returns
+        -------
+        TermReason
+            Value indicating reason this event would terminate integration
+        str
+            Extra information associated with reason for termination.
+
+        Notes
+        -----
+        The value `TermReason.UNKNOWN` should be used for reasons that do
+        not have an existing reason. For these termination situations the
+        message can provide more details.
+        """
+        return TermReason.TRANSITIONED, ""
 
 
 class TestEdgeVelocity(unittest.TestCase):
     """Class to test various functions and curve fits for Thwaites method"""
 
     # define the edge velocity functions
-    @classmethod
-    def U_e_fun(cls, x, C, m):
-        """Return edge velocity."""
+    @staticmethod
+    def u_e_fun(x: InputParam, u_ref: float, m: float) -> npt.NDArray:
+        """
+        Return edge velocity.
+
+        Parameters
+        ----------
+        x : InputParam
+            Streamwise location.
+        u_ref : float
+            Scale for edge velocity model.
+        m : float
+            Exponent for edge velocity model.
+
+        Returns
+        -------
+        numpy.ndarray
+            Edge velocity.
+        """
         x = np.asarray(x)
         if m == 0:
-            return C*np.ones_like(x)
-        return C*x**m
+            return u_ref*np.ones_like(x)
+        return u_ref*x**m
 
-    @classmethod
-    def dU_edx_fun(cls, x, C, m):
-        """Return the streamwise derivative of edge velocity."""
+    @staticmethod
+    def du_e_fun(x: InputParam, u_ref: float, m: float) -> npt.NDArray:
+        """
+        Return the streamwise derivative of edge velocity.
+
+        Parameters
+        ----------
+        x : InputParam
+            Streamwise location.
+        u_ref : float
+            Scale for edge velocity model.
+        m : float
+            Exponent for edge velocity model.
+
+        Returns
+        -------
+        numpy.ndarray
+            First derivative of edge velocity.
+        """
         x = np.asarray(x)
         if m == 0:
             return np.zeros_like(x)
         if m == 1:
-            return C*np.ones_like(x)
-        return m*C*x**(m-1)
+            return u_ref*np.ones_like(x)
+        return m*u_ref*x**(m-1)
 
-    @classmethod
-    def d2U_edx2_fun(cls, x, C, m):
-        """Return the streamwise second derivative of edge velocity."""
+    @staticmethod
+    def d2u_e_fun(x: InputParam, u_ref: float, m: float) -> npt.NDArray:
+        """
+        Return the streamwise second derivative of edge velocity.
+
+        Parameters
+        ----------
+        x : InputParam
+            Streamwise location.
+        u_ref : float
+            Scale for edge velocity model.
+        m : float
+            Exponent for edge velocity model.
+
+        Returns
+        -------
+        numpy.ndarray
+            Second derivative of edge velocity.
+        """
         x = np.asarray(x)
         if m in (0, 1):
             return np.zeros_like(x)
         if m == 2:
-            return m*C*np.ones_like(x)
-        return m*(m-1)*C*x**(m-2)
+            return m*u_ref*np.ones_like(x)
+        return m*(m-1)*u_ref*x**(m-2)
 
-    @classmethod
-    def d3U_edx3_fun(cls, x, C, m):
-        """Return the streamwise third derivative of edge velocity."""
+    @staticmethod
+    def d3u_e_fun(x: InputParam, u_ref: float, m: float) -> npt.NDArray:
+        """
+        Return the streamwise third derivative of edge velocity.
+
+        Parameters
+        ----------
+        x : InputParam
+            Streamwise location.
+        u_ref : float
+            Scale for edge velocity model.
+        m : float
+            Exponent for edge velocity model.
+
+        Returns
+        -------
+        numpy.ndarray
+            Third derivative of edge velocity.
+        """
         x = np.asarray(x)
         if m in (0, 1, 2):
             return np.zeros_like(x)
         if m == 3:
-            return m*(m-1)*C*np.ones_like(x)
-        return m*(m-1)*(m-2)*C*x**(m-3)
+            return m*(m-1)*u_ref*np.ones_like(x)
+        return m*(m-1)*(m-2)*u_ref*x**(m-3)
 
-    def test_setting_velocity_functions(self):
+    def test_setting_velocity_functions(self) -> None:
         """Test setting the velocity functions."""
         # create test class with all three functions
-        U_inf = 10
+        u_inf = 10
         m = 0.75
-        iblb = IBLMethodTest(U_e=lambda x: self.U_e_fun(x, U_inf, m),
-                             dU_edx=lambda x: self.dU_edx_fun(x, U_inf, m),
-                             d2U_edx2=lambda x: self.d2U_edx2_fun(x, U_inf, m))
+        def u_e_fun(x: InputParam) -> InputParam:
+            return self.u_e_fun(x, u_inf, m)
+
+
+        def du_e_fun(x: InputParam) -> InputParam:
+            return self.du_e_fun(x, u_inf, m)
+
+
+        def d2u_e_fun(x: InputParam) -> InputParam:
+            return self.d2u_e_fun(x, u_inf, m)
+
+
+        iblb = IBLMethodTest(u_e=u_e_fun, du_e=du_e_fun, d2u_e=d2u_e_fun)
 
         x = np.linspace(0.1, 5, 21)
-        U_e_ref = self.U_e_fun(x, U_inf, m)
-        dU_edx_ref = self.dU_edx_fun(x, U_inf, m)
-        d2U_edx2_ref = self.d2U_edx2_fun(x, U_inf, m)
-        self.assertIsNone(npt.assert_allclose(iblb.U_e(x), U_e_ref))
-        self.assertIsNone(npt.assert_allclose(iblb.dU_edx(x), dU_edx_ref))
-        self.assertIsNone(npt.assert_allclose(iblb.d2U_edx2(x), d2U_edx2_ref))
+        u_e_ref = self.u_e_fun(x, u_inf, m)
+        du_e_ref = self.du_e_fun(x, u_inf, m)
+        d2u_e_ref = self.d2u_e_fun(x, u_inf, m)
+        self.assertIsNone(np_type.assert_allclose(iblb.u_e(x), u_e_ref))
+        self.assertIsNone(np_type.assert_allclose(iblb.du_e(x), du_e_ref))
+        self.assertIsNone(np_type.assert_allclose(iblb.d2u_e(x), d2u_e_ref))
 
         # create test class with two functions
-        U_inf = 10
+        u_inf = 10
         m = 0.75
-        iblb = IBLMethodTest(U_e=lambda x: self.U_e_fun(x, U_inf, m),
-                             dU_edx=lambda x: self.dU_edx_fun(x, U_inf, m))
+        iblb = IBLMethodTest(u_e=u_e_fun, du_e=du_e_fun)
 
         x = np.linspace(0.1, 5, 21)
-        U_e_ref = self.U_e_fun(x, U_inf, m)
-        dU_edx_ref = self.dU_edx_fun(x, U_inf, m)
-        d2U_edx2_ref = self.d2U_edx2_fun(x, U_inf, m)
-        self.assertIsNone(npt.assert_allclose(iblb.U_e(x), U_e_ref))
-        self.assertIsNone(npt.assert_allclose(iblb.dU_edx(x), dU_edx_ref))
-        self.assertIsNone(npt.assert_allclose(iblb.d2U_edx2(x), d2U_edx2_ref))
+        u_e_ref = self.u_e_fun(x, u_inf, m)
+        du_e_ref = self.du_e_fun(x, u_inf, m)
+        d2u_e_ref = self.d2u_e_fun(x, u_inf, m)
+        self.assertIsNone(np_type.assert_allclose(iblb.u_e(x), u_e_ref))
+        self.assertIsNone(np_type.assert_allclose(iblb.du_e(x), du_e_ref))
+        self.assertIsNone(np_type.assert_allclose(iblb.d2u_e(x), d2u_e_ref))
 
         # create test class with one function
-        U_inf = 10
+        u_inf = 10
         m = 0.75
-        iblb = IBLMethodTest(U_e=lambda x: self.U_e_fun(x, U_inf, m))
+        iblb = IBLMethodTest(u_e=u_e_fun)
 
         x = np.linspace(0.1, 5, 21)
-        U_e_ref = self.U_e_fun(x, U_inf, m)
-        dU_edx_ref = self.dU_edx_fun(x, U_inf, m)
-        d2U_edx2_ref = self.d2U_edx2_fun(x, U_inf, m)
-        self.assertIsNone(npt.assert_allclose(iblb.U_e(x), U_e_ref))
-        self.assertIsNone(npt.assert_allclose(iblb.dU_edx(x), dU_edx_ref))
+        u_e_ref = self.u_e_fun(x, u_inf, m)
+        du_e_ref = self.du_e_fun(x, u_inf, m)
+        d2u_e_ref = self.d2u_e_fun(x, u_inf, m)
+        self.assertIsNone(np_type.assert_allclose(iblb.u_e(x), u_e_ref))
+        self.assertIsNone(np_type.assert_allclose(iblb.du_e(x), du_e_ref))
         # NOTE: second derivative has slightly larger errors
-        self.assertIsNone(npt.assert_allclose(iblb.d2U_edx2(x), d2U_edx2_ref,
-                                              rtol=1e-5, atol=0))
+        self.assertIsNone(np_type.assert_allclose(iblb.d2u_e(x), d2u_e_ref,
+                                                  rtol=1e-5, atol=0))
 
-    def test_setting_velocity_splines(self):
+    def test_setting_velocity_splines(self) -> None:
         """Test setting the velocity with splines."""
         # set the edge velocity spline
         x_sample = np.linspace(0.1, 5, 8)
-        U_inf = 10
+        u_inf = 10
         m = 1.25
-        U_e = PchipInterpolator(x_sample, self.U_e_fun(x_sample, U_inf, m))
-        dU_edx = U_e.derivative()
-        d2U_edx2 = dU_edx.derivative()
-        iblb = IBLMethodTest(U_e=U_e)
+        u_e = PchipInterpolator(x_sample, self.u_e_fun(x_sample, u_inf, m))
+        du_e = u_e.derivative()
+        d2u_e = du_e.derivative()
+        iblb = IBLMethodTest(u_e=u_e)
 
         x = np.linspace(0.1, 5, 21)
-        U_e_ref = U_e(x)
-        dU_edx_ref = dU_edx(x)
-        d2U_edx2_ref = d2U_edx2(x)
-        self.assertIsNone(npt.assert_allclose(iblb.U_e(x), U_e_ref))
-        self.assertIsNone(npt.assert_allclose(iblb.dU_edx(x), dU_edx_ref))
-        self.assertIsNone(npt.assert_allclose(iblb.d2U_edx2(x), d2U_edx2_ref))
+        u_e_ref = u_e(x)
+        du_e_ref = du_e(x)
+        d2u_e_ref = d2u_e(x)
+        self.assertIsNone(np_type.assert_allclose(iblb.u_e(x), u_e_ref))
+        self.assertIsNone(np_type.assert_allclose(iblb.du_e(x), du_e_ref))
+        self.assertIsNone(np_type.assert_allclose(iblb.d2u_e(x), d2u_e_ref))
 
         # set the edge velocity derivative spline
         x_sample = np.linspace(0.1, 5, 8)
-        U_inf = 10
+        u_inf = 10
         m = 1.25
-        dU_edx = PchipInterpolator(x_sample,
-                                   self.dU_edx_fun(x_sample, U_inf, m))
-        U_e = dU_edx.antiderivative()
-        U_e.c[-1,:] = U_e.c[-1,:]+self.U_e_fun(x_sample[0], U_inf, m)
-        d2U_edx2 = dU_edx.derivative()
-        iblb = IBLMethodTest(U_e=U_e, dU_edx=dU_edx)
+        du_e = PchipInterpolator(x_sample, self.du_e_fun(x_sample, u_inf, m))
+        u_e = du_e.antiderivative()
+        u_e.c[-1,:] = u_e.c[-1,:]+self.u_e_fun(x_sample[0], u_inf, m)
+        d2u_e = du_e.derivative()
+        iblb = IBLMethodTest(u_e=u_e, du_e=du_e)
 
         x = np.linspace(0.1, 5, 21)
-        U_e_ref = U_e(x)
-        dU_edx_ref = dU_edx(x)
-        d2U_edx2_ref = d2U_edx2(x)
-        self.assertIsNone(npt.assert_allclose(iblb.U_e(x), U_e_ref))
-        self.assertIsNone(npt.assert_allclose(iblb.dU_edx(x), dU_edx_ref))
-        self.assertIsNone(npt.assert_allclose(iblb.d2U_edx2(x), d2U_edx2_ref))
+        u_e_ref = u_e(x)
+        du_e_ref = du_e(x)
+        d2u_e_ref = d2u_e(x)
+        self.assertIsNone(np_type.assert_allclose(iblb.u_e(x), u_e_ref))
+        self.assertIsNone(np_type.assert_allclose(iblb.du_e(x), du_e_ref))
+        self.assertIsNone(np_type.assert_allclose(iblb.d2u_e(x), d2u_e_ref))
 
         # set the edge velocity second derivative spline
         x_sample = np.linspace(0.1, 5, 8)
-        U_inf = 10
+        u_inf = 10
         m = 1.25
-        d2U_edx2 = PchipInterpolator(x_sample,
-                                     self.d2U_edx2_fun(x_sample, U_inf, m))
-        dU_edx = d2U_edx2.antiderivative()
-        dU_edx.c[-1,:] = dU_edx.c[-1,:]+self.dU_edx_fun(x_sample[0], U_inf, m)
-        U_e = dU_edx.antiderivative()
-        U_e.c[-1,:] = U_e.c[-1,:]+self.U_e_fun(x_sample[0], U_inf, m)
-        iblb = IBLMethodTest(U_e=U_e, dU_edx=dU_edx, d2U_edx2=d2U_edx2)
+        d2u_e = PchipInterpolator(x_sample, self.d2u_e_fun(x_sample, u_inf, m))
+        du_e = d2u_e.antiderivative()
+        du_e.c[-1,:] = du_e.c[-1,:]+self.du_e_fun(x_sample[0], u_inf, m)
+        u_e = du_e.antiderivative()
+        u_e.c[-1,:] = u_e.c[-1,:]+self.u_e_fun(x_sample[0], u_inf, m)
+        iblb = IBLMethodTest(u_e=u_e, du_e=du_e, d2u_e=d2u_e)
 
         x = np.linspace(0.1, 5, 21)
-        U_e_ref = U_e(x)
-        dU_edx_ref = dU_edx(x)
-        d2U_edx2_ref = d2U_edx2(x)
-        self.assertIsNone(npt.assert_allclose(iblb.U_e(x), U_e_ref))
-        self.assertIsNone(npt.assert_allclose(iblb.dU_edx(x), dU_edx_ref))
-        self.assertIsNone(npt.assert_allclose(iblb.d2U_edx2(x), d2U_edx2_ref))
+        u_e_ref = u_e(x)
+        du_e_ref = du_e(x)
+        d2u_e_ref = d2u_e(x)
+        self.assertIsNone(np_type.assert_allclose(iblb.u_e(x), u_e_ref))
+        self.assertIsNone(np_type.assert_allclose(iblb.du_e(x), du_e_ref))
+        self.assertIsNone(np_type.assert_allclose(iblb.d2u_e(x), d2u_e_ref))
 
-    def test_setting_velocity_points(self):
+    def test_setting_velocity_points(self) -> None:
         """Test setting velocity from points."""
         # set the edge velocity values
         x_sample = np.linspace(0.1, 5, 8)
-        U_inf = 10
+        u_inf = 10
         m = 1.25
-        U_e = [x_sample, self.U_e_fun(x_sample, U_inf, m)]
-        U_e_spline = PchipInterpolator(x_sample, self.U_e_fun(x_sample,
-                                                              U_inf, m))
-        dU_edx_spline = U_e_spline.derivative()
-        d2U_edx2_spline = dU_edx_spline.derivative()
-        iblb = IBLMethodTest(U_e=U_e)
+        u_e = [x_sample, self.u_e_fun(x_sample, u_inf, m)]
+        u_e_spline = PchipInterpolator(x_sample, self.u_e_fun(x_sample,
+                                                              u_inf, m))
+        du_e_spline = u_e_spline.derivative()
+        d2u_e_spline = du_e_spline.derivative()
+        iblb = IBLMethodTest(u_e=u_e)
 
         x = np.linspace(0.1, 5, 21)
-        U_e_ref = U_e_spline(x)
-        dU_edx_ref = dU_edx_spline(x)
-        d2U_edx2_ref = d2U_edx2_spline(x)
-        self.assertIsNone(npt.assert_allclose(iblb.U_e(x), U_e_ref))
-        self.assertIsNone(npt.assert_allclose(iblb.dU_edx(x), dU_edx_ref))
-        self.assertIsNone(npt.assert_allclose(iblb.d2U_edx2(x), d2U_edx2_ref))
+        u_e_ref = u_e_spline(x)
+        du_e_ref = du_e_spline(x)
+        d2u_e_ref = d2u_e_spline(x)
+        self.assertIsNone(np_type.assert_allclose(iblb.u_e(x), u_e_ref))
+        self.assertIsNone(np_type.assert_allclose(iblb.du_e(x), du_e_ref))
+        self.assertIsNone(np_type.assert_allclose(iblb.d2u_e(x), d2u_e_ref))
 
+    def test_setting_velocity_derivative_points(self) -> None:
+        """Test setting velocity derivative from points."""
         # set the edge velocity derivative points
         x_sample = np.linspace(0.1, 5, 8)
-        U_inf = 10
+        u_inf = 10
         m = 1.25
-        U_e = self.U_e_fun(x_sample[0], U_inf, m)
-        dU_edx = [x_sample, self.dU_edx_fun(x_sample, U_inf, m)]
-        dU_edx_spline = PchipInterpolator(x_sample, self.dU_edx_fun(x_sample,
-                                                                    U_inf, m))
-        U_e_spline = dU_edx_spline.antiderivative()
-        U_e_spline.c[-1,:] = (U_e_spline.c[-1,:]
-                              + self.U_e_fun(x_sample[0], U_inf, m))
-        d2U_edx2_spline = dU_edx_spline.derivative()
-        iblb = IBLMethodTest(U_e=U_e, dU_edx=dU_edx)
+        u_e = self.u_e_fun(x_sample[0], u_inf, m)
+        du_e = [x_sample, self.du_e_fun(x_sample, u_inf, m)]
+        du_e_spline = PchipInterpolator(x_sample, self.du_e_fun(x_sample,
+                                                                u_inf, m))
+        u_e_spline = du_e_spline.antiderivative()
+        u_e_spline.c[-1,:] = (u_e_spline.c[-1,:]
+                              + self.u_e_fun(x_sample[0], u_inf, m))
+        d2u_e_spline = du_e_spline.derivative()
+        iblb = IBLMethodTest(u_e=u_e, du_e=du_e)
 
         x = np.linspace(0.1, 5, 21)
-        U_e_ref = U_e_spline(x)
-        dU_edx_ref = dU_edx_spline(x)
-        d2U_edx2_ref = d2U_edx2_spline(x)
-        self.assertIsNone(npt.assert_allclose(iblb.U_e(x), U_e_ref))
-        self.assertIsNone(npt.assert_allclose(iblb.dU_edx(x), dU_edx_ref))
-        self.assertIsNone(npt.assert_allclose(iblb.d2U_edx2(x), d2U_edx2_ref))
+        u_e_ref = u_e_spline(x)
+        du_e_ref = du_e_spline(x)
+        d2u_e_ref = d2u_e_spline(x)
+        self.assertIsNone(np_type.assert_allclose(iblb.u_e(x), u_e_ref))
+        self.assertIsNone(np_type.assert_allclose(iblb.du_e(x), du_e_ref))
+        self.assertIsNone(np_type.assert_allclose(iblb.d2u_e(x), d2u_e_ref))
 
-    def test_delay_setting_velocity(self):
+    def test_delay_setting_velocity(self) -> None:
         """Test setting the velocity after class creation."""
         # create test class with all three functions
-        U_inf = 10
+        u_inf = 10
         m = 0.75
         iblb = IBLMethodTest()
         x = np.linspace(0.1, 5, 21)
-        U_e_ref = self.U_e_fun(x, U_inf, m)
-        dU_edx_ref = self.dU_edx_fun(x, U_inf, m)
-        d2U_edx2_ref = self.d2U_edx2_fun(x, U_inf, m)
+
+        def u_e_fun(x: InputParam) -> InputParam:
+            return self.u_e_fun(x, u_inf, m)
+
+
+        def du_e_fun(x: InputParam) -> InputParam:
+            return self.du_e_fun(x, u_inf, m)
+
+
+        def d2u_e_fun(x: InputParam) -> InputParam:
+            return self.d2u_e_fun(x, u_inf, m)
+
+
+        u_e_ref = u_e_fun(x)
+        du_e_ref = du_e_fun(x)
+        d2u_e_ref = d2u_e_fun(x)
 
         with self.assertRaises(ValueError):
-            iblb.U_e(x)
+            _ = iblb.u_e(x)
         with self.assertRaises(ValueError):
-            iblb.dU_edx(x)
+            _ = iblb.du_e(x)
         with self.assertRaises(ValueError):
-            iblb.d2U_edx2(x)
+            _ = iblb.d2u_e(x)
 
-        iblb.set_velocity(U_e=lambda x: self.U_e_fun(x, U_inf, m),
-                          dU_edx=lambda x: self.dU_edx_fun(x, U_inf, m),
-                          d2U_edx2=lambda x: self.d2U_edx2_fun(x, U_inf, m))
-        self.assertIsNone(npt.assert_allclose(iblb.U_e(x), U_e_ref))
-        self.assertIsNone(npt.assert_allclose(iblb.dU_edx(x), dU_edx_ref))
-        self.assertIsNone(npt.assert_allclose(iblb.d2U_edx2(x), d2U_edx2_ref))
+        iblb.set_velocity(u_e=u_e_fun, du_e=du_e_fun, d2u_e=d2u_e_fun)
+        self.assertIsNone(np_type.assert_allclose(iblb.u_e(x), u_e_ref))
+        self.assertIsNone(np_type.assert_allclose(iblb.du_e(x), du_e_ref))
+        self.assertIsNone(np_type.assert_allclose(iblb.d2u_e(x), d2u_e_ref))
 
-    def test_terminating_solver(self):
+    def test_terminating_solver(self) -> None:
         """Test early termination of solver."""
-        U_inf = 10
+        u_inf = 10
         m = 1
         x_kill = 3
-        iblb = IBLMethodTest(U_e=lambda x: self.U_e_fun(x, U_inf, m),
-                             x_kill=x_kill)
+
+        def u_e_fun(x: InputParam) -> InputParam:
+            return self.u_e_fun(x, u_inf, m)
+
+
+        iblb = IBLMethodTest(u_e=u_e_fun, x_kill=x_kill)
+
+        # test setting viscosity
+        iblb.nu = 1e-5
+        self.assertAlmostEqual(iblb.nu, 1e-5)
+        with self.assertRaises(ValueError):
+            iblb.nu = 0.0
 
         # go through the entire xrange
         #
         # NOTE: No need to set the velocity terms because they are not used in
         #       this basic implementation.
         # NOTE: This solves the simple differential equation y'=x
-        def ref_fun(x):
+        def ref_fun(x: InputParam) -> npt.NDArray:
             return np.array([0.5*x**2+1])
 
         x_start = 1
@@ -389,7 +680,7 @@ class TestEdgeVelocity(unittest.TestCase):
         self.assertEqual(rtn.status, 0)
         self.assertEqual(rtn.message, "Completed")
         self.assertEqual(rtn.x_end, x_end)
-        self.assertIsNone(npt.assert_allclose(rtn.F_end, ref_fun(x_end)))
+        self.assertIsNone(np_type.assert_allclose(rtn.f_end, ref_fun(x_end)))
 
         # stop because solver terminated early
         x_start = 1
@@ -397,10 +688,10 @@ class TestEdgeVelocity(unittest.TestCase):
         iblb.y0 = ref_fun(x_start)
         rtn = iblb.solve(x_start, x_end)
         self.assertTrue(rtn.success)
-        self.assertEqual(rtn.status, -1)
+        self.assertEqual(rtn.status, TermReason.SEPARATED)
         self.assertEqual(rtn.message, "Separated")
         self.assertEqual(rtn.x_end, x_kill)
-        self.assertIsNone(npt.assert_allclose(rtn.F_end, ref_fun(x_kill)))
+        self.assertIsNone(np_type.assert_allclose(rtn.f_end, ref_fun(x_kill)))
 
         # stop because external trigger
         x_start = 1
@@ -411,11 +702,11 @@ class TestEdgeVelocity(unittest.TestCase):
         rtn = iblb.solve(x_start, x_end,
                          term_event=IBLMethodTestTransition(y_trans))
         self.assertTrue(rtn.success)
-        self.assertEqual(rtn.status, 1)
+        self.assertEqual(rtn.status, TermReason.TRANSITIONED)
         self.assertEqual(rtn.message, "Transition")
         self.assertEqual(rtn.x_end, x_trans)
-        self.assertIsNone(npt.assert_allclose(rtn.F_end, y_trans))
+        self.assertIsNone(np_type.assert_allclose(rtn.f_end, y_trans))
 
 
 if __name__ == "__main__":
-    unittest.main(verbosity=1)
+    _ = unittest.main(verbosity=1)
